@@ -1,9 +1,5 @@
-﻿using System.Threading.Tasks;
-
-namespace EventBus.Abstractions
+﻿namespace EventBus.Abstractions;
+public interface IDynamicIntegrationEventHandler
 {
-    public interface IDynamicIntegrationEventHandler
-    {
-        Task Handle(dynamic eventData);
-    }
+    Task Handle(dynamic eventData);
 }
