@@ -1,151 +1,143 @@
-﻿using EventBus.Abstractions;
-using EventBus.Events;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-
-namespace EventBus
+﻿namespace EventBus;
+public partial class InMemoryEventBusSubscriptionsManager : IEventBusSubscriptionsManager
 {
-    public partial class InMemoryEventBusSubscriptionsManager : IEventBusSubscriptionsManager
+    private readonly Dictionary<string, List<SubscriptionInfo>> _handlers;
+    private readonly List<Type> _eventTypes;
+
+    public event EventHandler<string> OnEventRemoved;
+
+    public InMemoryEventBusSubscriptionsManager()
     {
-        private readonly Dictionary<string, List<SubscriptionInfo>> _handlers;
-        private readonly List<Type> _eventTypes;
+        _handlers = new Dictionary<string, List<SubscriptionInfo>>();
+        _eventTypes = new List<Type>();
+    }
 
-        public event EventHandler<string> OnEventRemoved;
+    public bool IsEmpty => !_handlers.Keys.Any();
+    public void Clear() => _handlers.Clear();
 
-        public InMemoryEventBusSubscriptionsManager()
+    public void AddDynamicSubScription<TH>(string eventName)
+        where TH : IDynamicIntegrationEventHandler
+    {
+        DoAddSubscription(typeof(TH), eventName, true);
+    }
+
+    public void AddSubsciption<T, TH>()
+        where T : IntegrationEvent
+        where TH : IIntegrationEventHandler<T>
+    {
+        var eventName = GetEventKey<T>();
+
+        DoAddSubscription(typeof(TH), eventName, false);
+
+        if (!_eventTypes.Contains(typeof(T)))
         {
-            _handlers = new Dictionary<string, List<SubscriptionInfo>>();
-            _eventTypes = new List<Type>();
+            _eventTypes.Add(typeof(T));
+        }
+    }
+
+    public string GetEventKey<T>()
+    {
+        return typeof(T).Name;
+    }
+
+    public Type GetEventTypeByName(string eventName) => _eventTypes.SingleOrDefault(t => t.Name == eventName);
+
+    public IEnumerable<SubscriptionInfo> GetHandlersForEvent<T>() where T : IntegrationEvent
+    {
+        var key = GetEventKey<T>();
+        return GetHandlersForEvent(key);
+    }
+
+    public IEnumerable<SubscriptionInfo> GetHandlersForEvent(string eventName) => _handlers[eventName];
+
+    public bool HasSubscriptionsForEvent<T>() where T : IntegrationEvent
+    {
+        var key = GetEventKey<T>();
+        return HasSubscriptionsForEvent(key);
+    }
+
+    public bool HasSubscriptionsForEvent(string eventName) => _handlers.ContainsKey(eventName);
+
+    public void RemoveDynamicSubscription<TH>(string eventName)
+        where TH : IDynamicIntegrationEventHandler
+    {
+        var handlerToRemove = FindDynamicSubscriptionToRemove<TH>(eventName);
+        DoRemoveHandler(eventName, handlerToRemove);
+    }
+
+    public void RemoveSubscription<T, TH>()
+        where T : IntegrationEvent
+        where TH : IIntegrationEventHandler<T>
+    {
+        var handlerToRemove = FindSubscriptionToRemove<T, TH>();
+        var eventName = GetEventKey<T>();
+        DoRemoveHandler(eventName, handlerToRemove);
+    }
+
+    private void DoAddSubscription(Type handlerType, string eventName, bool isDynamic)
+    {
+        if (!HasSubscriptionsForEvent(eventName))
+        {
+            _handlers.Add(eventName, new List<SubscriptionInfo>());
         }
 
-        public bool IsEmpty => !_handlers.Keys.Any();
-        public void Clear() => _handlers.Clear();
-
-        public void AddDynamicSubScription<TH>(string eventName)
-            where TH : IDynamicIntegrationEventHandler
+        if (_handlers[eventName].Any(s => s.HandlerType == handlerType))
         {
-            DoAddSubscription(typeof(TH), eventName, true);
+            throw new ArgumentException($"Handler Type {handlerType.Name} already registered for '{eventName}'", nameof(handlerType));
         }
-
-        public void AddSubsciption<T, TH>()
-            where T : IntegrationEvent
-            where TH : IIntegrationEventHandler<T>
+        if (isDynamic)
         {
-            var eventName = GetEventKey<T>();
+            _handlers[eventName].Add(SubscriptionInfo.Dynamic(handlerType));
+        }
+        else
+        {
+            _handlers[eventName].Add(SubscriptionInfo.Typed(handlerType));
+        }
+    }
 
-            DoAddSubscription(typeof(TH), eventName, false);
-
-            if (!_eventTypes.Contains(typeof(T)))
+    private void DoRemoveHandler(string eventName, SubscriptionInfo subscriptionToRemove)
+    {
+        if (subscriptionToRemove != null)
+        {
+            _handlers[eventName].Remove(subscriptionToRemove);
+            if (!_handlers[eventName].Any())
             {
-                _eventTypes.Add(typeof(T));
-            }
-        }
-
-        public string GetEventKey<T>()
-        {
-            return typeof(T).Name;
-        }
-
-        public Type GetEventTypeByName(string eventName) => _eventTypes.SingleOrDefault(t => t.Name == eventName);
-
-        public IEnumerable<SubscriptionInfo> GetHandlersForEvent<T>() where T : IntegrationEvent
-        {
-            var key = GetEventKey<T>();
-            return GetHandlersForEvent(key);
-        }
-
-        public IEnumerable<SubscriptionInfo> GetHandlersForEvent(string eventName) => _handlers[eventName];
-
-        public bool HasSubscriptionsForEvent<T>() where T : IntegrationEvent
-        {
-            var key = GetEventKey<T>();
-            return HasSubscriptionsForEvent(key);
-        }
-
-        public bool HasSubscriptionsForEvent(string eventName) => _handlers.ContainsKey(eventName);
-
-        public void RemoveDynamicSubscription<TH>(string eventName)
-            where TH : IDynamicIntegrationEventHandler
-        {
-            var handlerToRemove = FindDynamicSubscriptionToRemove<TH>(eventName);
-            DoRemoveHandler(eventName, handlerToRemove);
-        }
-
-        public void RemoveSubscription<T, TH>()
-            where T : IntegrationEvent
-            where TH : IIntegrationEventHandler<T>
-        {
-            var handlerToRemove = FindSubscriptionToRemove<T, TH>();
-            var eventName = GetEventKey<T>();
-            DoRemoveHandler(eventName, handlerToRemove);
-        }
-
-        private void DoAddSubscription(Type handlerType, string eventName, bool isDynamic)
-        {
-            if (!HasSubscriptionsForEvent(eventName))
-            {
-                _handlers.Add(eventName, new List<SubscriptionInfo>());
-            }
-
-            if (_handlers[eventName].Any(s => s.HandlerType == handlerType))
-            {
-                throw new ArgumentException($"Handler Type {handlerType.Name} already registered for '{eventName}'", nameof(handlerType));
-            }
-            if (isDynamic)
-            {
-                _handlers[eventName].Add(SubscriptionInfo.Dynamic(handlerType));
-            }
-            else
-            {
-                _handlers[eventName].Add(SubscriptionInfo.Typed(handlerType));
-            }
-        }
-
-        private void DoRemoveHandler(string eventName, SubscriptionInfo subscriptionToRemove)
-        {
-            if (subscriptionToRemove != null)
-            {
-                _handlers[eventName].Remove(subscriptionToRemove);
-                if (!_handlers[eventName].Any())
+                _handlers.Remove(eventName);
+                var eventType = _eventTypes.SingleOrDefault(e => e.Name == eventName);
+                if (eventType != null)
                 {
-                    _handlers.Remove(eventName);
-                    var eventType = _eventTypes.SingleOrDefault(e => e.Name == eventName);
-                    if (eventType != null)
-                    {
-                        _eventTypes.Remove(eventType);
-                    }
-                    RaiseOnEventRemoved(eventName);
+                    _eventTypes.Remove(eventType);
                 }
+                RaiseOnEventRemoved(eventName);
             }
         }
-        private void RaiseOnEventRemoved(string eventName)
-        {
-            var handler = OnEventRemoved;
-            handler?.Invoke(this, eventName);
-        }
+    }
+    private void RaiseOnEventRemoved(string eventName)
+    {
+        var handler = OnEventRemoved;
+        handler?.Invoke(this, eventName);
+    }
 
-        private SubscriptionInfo FindDynamicSubscriptionToRemove<TH>(string eventName)
-            where TH : IDynamicIntegrationEventHandler
-        {
-            return DoFindSubscriptionToRemove(eventName, typeof(TH));
-        }
+    private SubscriptionInfo FindDynamicSubscriptionToRemove<TH>(string eventName)
+        where TH : IDynamicIntegrationEventHandler
+    {
+        return DoFindSubscriptionToRemove(eventName, typeof(TH));
+    }
 
-        private SubscriptionInfo FindSubscriptionToRemove<T, TH>()
-            where T : IntegrationEvent
-            where TH : IIntegrationEventHandler<T>
-        {
-            var eventName = GetEventKey<T>();
-            return DoFindSubscriptionToRemove(eventName, typeof(TH));
-        }
+    private SubscriptionInfo FindSubscriptionToRemove<T, TH>()
+        where T : IntegrationEvent
+        where TH : IIntegrationEventHandler<T>
+    {
+        var eventName = GetEventKey<T>();
+        return DoFindSubscriptionToRemove(eventName, typeof(TH));
+    }
 
-        private SubscriptionInfo DoFindSubscriptionToRemove(string eventName, Type handlerType)
+    private SubscriptionInfo DoFindSubscriptionToRemove(string eventName, Type handlerType)
+    {
+        if (!HasSubscriptionsForEvent(eventName))
         {
-            if (!HasSubscriptionsForEvent(eventName))
-            {
-                return null;
-            }
-            return _handlers[eventName].SingleOrDefault(s => s.HandlerType == handlerType);
+            return null;
         }
+        return _handlers[eventName].SingleOrDefault(s => s.HandlerType == handlerType);
     }
 }
