@@ -1,263 +1,245 @@
-﻿using EventBus.Abstractions;
-using EventBus.Events;
-using Microsoft.Extensions.Logging;
-using System;
-using EventBus;
-using RabbitMQ.Client;
-using Autofac;
-using RabbitMQ.Client.Events;
-using System.Threading.Tasks;
-using System.Text;
-using System.Text.Json;
-using Polly.Retry;
-using RabbitMQ.Client.Exceptions;
-using System.Net.Sockets;
-using Polly;
-
-namespace EventBusRabbitMQ
+﻿namespace EventBusRabbitMQ;
+public class EventBusRabbitMQ : IEventBus, IDisposable
 {
-    public class EventBusRabbitMQ : IEventBus, IDisposable
+    const string BROKER_NAME = "eshop_event_bus";
+    const string AUTOFAC_SCOPE_NAME = "eshop_event_bus";
+
+    private readonly IRabbitMQPersistentConnection _persistentConnection;
+    private readonly ILogger<EventBusRabbitMQ> _logger;
+    private readonly IEventBusSubscriptionsManager _subscriptionsManager;
+    private readonly ILifetimeScope _autofac;
+    private readonly int _retryCount;
+
+    private IModel _consumerChannel;
+    private string _queueName;
+
+    public EventBusRabbitMQ(
+        IRabbitMQPersistentConnection persistentConnection,
+        ILogger<EventBusRabbitMQ> logger, ILifetimeScope autofac,
+        IEventBusSubscriptionsManager subscriptionsManager, string queueName = null,
+        int retryCount = 5)
     {
-        const string BROKER_NAME = "eshop_event_bus";
-        const string AUTOFAC_SCOPE_NAME = "eshop_event_bus";
-
-        private readonly IRabbitMQPersistentConnection _persistentConnection;
-        private readonly ILogger<EventBusRabbitMQ> _logger;
-        private readonly IEventBusSubscriptionsManager _subscriptionsManager;
-        private readonly ILifetimeScope _autofac;
-        private readonly int _retryCount;
-
-        private IModel _consumerChannel;
-        private string _queueName;
-
-        public EventBusRabbitMQ(
-            IRabbitMQPersistentConnection persistentConnection,
-            ILogger<EventBusRabbitMQ> logger, ILifetimeScope autofac,
-            IEventBusSubscriptionsManager subscriptionsManager, string queueName = null,
-            int retryCount = 5)
+        _persistentConnection = persistentConnection ?? throw new AccessViolationException(nameof(persistentConnection));
+        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        _subscriptionsManager = subscriptionsManager ?? throw new ArgumentNullException(nameof(subscriptionsManager));
+        _queueName = queueName;
+        _autofac = autofac;
+        _retryCount = retryCount;
+        _consumerChannel = CreateConsumerChannel();
+        _subscriptionsManager.OnEventRemoved += SubscriptionsManager_OnEventRemoved;
+    }
+    public void Publish(IntegrationEvent @event)
+    {
+        if (!_persistentConnection.IsConnected)
         {
-            _persistentConnection = persistentConnection ?? throw new AccessViolationException(nameof(persistentConnection));
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _subscriptionsManager = subscriptionsManager ?? throw new ArgumentNullException(nameof(subscriptionsManager));
-            _queueName = queueName;
-            _autofac = autofac;
-            _retryCount = retryCount;
-            _consumerChannel = CreateConsumerChannel();
-            _subscriptionsManager.OnEventRemoved += SubscriptionsManager_OnEventRemoved;
+            _persistentConnection.TryConnect();
         }
-        public void Publish(IntegrationEvent @event)
-        {
-            if (!_persistentConnection.IsConnected)
+        var policy = RetryPolicy.Handle<BrokerUnreachableException>()
+            .Or<SocketException>()
+            .WaitAndRetry(_retryCount, retryAttempt => TimeSpan.FromSeconds(Math.Pow(2, retryAttempt)), (ex, time) =>
             {
-                _persistentConnection.TryConnect();
-            }
-            var policy = RetryPolicy.Handle<BrokerUnreachableException>()
-                .Or<SocketException>()
-                .WaitAndRetry(_retryCount, retryAttempt => TimeSpan.FromSeconds(Math.Pow(2, retryAttempt)), (ex, time) =>
-                {
-                    _logger.LogWarning(ex, "Could not publish event: {EventId} after {Timeout}s ({ExceptionMessage})", @event.Id, $"{time.TotalSeconds}:n1", ex.Message);
-                });
+                _logger.LogWarning(ex, "Could not publish event: {EventId} after {Timeout}s ({ExceptionMessage})", @event.Id, $"{time.TotalSeconds}:n1", ex.Message);
+            });
 
-            var eventName = @event.GetType().Name;
-            _logger.LogTrace("Creating RabbitMQ channel to publish event: {EventId} ({EventName})", @event.Id, eventName);
+        var eventName = @event.GetType().Name;
+        _logger.LogTrace("Creating RabbitMQ channel to publish event: {EventId} ({EventName})", @event.Id, eventName);
 
-            using (var channel = _persistentConnection.CreateModel())
-            {
-                _logger.LogTrace("Declaring RabbitMQ exchange to publish event: {EventId}", @event.Id);
-
-                channel.ExchangeDeclare(exchange: BROKER_NAME, type: "direct");
-
-                var body = JsonSerializer.SerializeToUtf8Bytes(@event, @event.GetType(), new JsonSerializerOptions { WriteIndented = true });
-
-                policy.Execute(() =>
-                {
-                    var properties = channel.CreateBasicProperties();
-                    properties.DeliveryMode = 2;//persistent
-
-                    _logger.LogTrace("Publishing event to RabbitMQ: {EventId}", @event.Id);
-
-                    channel.BasicPublish(
-                        exchange: BROKER_NAME,
-                        routingKey: eventName,
-                        mandatory: true,
-                        basicProperties: properties,
-                        body: body);
-                });
-            }
-        }
-        public void Subscribe<T, TH>()
-            where T : IntegrationEvent
-            where TH : IIntegrationEventHandler<T>
+        using (var channel = _persistentConnection.CreateModel())
         {
-            var eventName = _subscriptionsManager.GetEventKey<T>();
-            DoInternalSubscription(eventName);
-
-            _logger.LogInformation("Subscribing to event {EventName} with {EventHandler}", eventName, typeof(TH));
-            _subscriptionsManager.AddSubsciption<T, TH>();
-            StartBasicConsume();
-        }
-        public void SubscribeDynamic<TH>(string eventName) where TH : IDynamicIntegrationEventHandler
-        {
-            _logger.LogInformation("Subscribing to dynamic event {EventName} with {EventHandler}", eventName, typeof(TH));
-            DoInternalSubscription(eventName);
-            _subscriptionsManager.AddDynamicSubScription<TH>(eventName);
-            StartBasicConsume();
-        }
-
-        public void Unsubscribe<T, TH>()
-            where T : IntegrationEvent
-            where TH : IIntegrationEventHandler<T>
-        {
-            var eventName = _subscriptionsManager.GetEventKey<T>();
-            _logger.LogInformation("Unsubscribing from event {EventName}", eventName);
-            _subscriptionsManager.RemoveSubscription<T, TH>();
-        }
-
-        public void UnsubscribeDynamic<TH>(string eventName) where TH : IDynamicIntegrationEventHandler
-        {
-            _logger.LogInformation("Unsubscribing from dynamic event {EventName}", eventName);
-            _subscriptionsManager.RemoveDynamicSubscription<TH>(eventName);
-        }
-        public void Dispose()
-        {
-            if (_consumerChannel != null)
-            {
-                _consumerChannel.Dispose();
-            }
-            _subscriptionsManager.Clear();
-        }
-
-        private void DoInternalSubscription(string eventName)
-        {
-            var containsKey = _subscriptionsManager.HasSubscriptionsForEvent(eventName);
-            if (!containsKey)
-            {
-                if (!_persistentConnection.IsConnected)
-                {
-                    _persistentConnection.TryConnect();
-                }
-                _consumerChannel.QueueBind(queue: _queueName, exchange: BROKER_NAME, routingKey: eventName);
-            }
-        }
-
-        private void SubscriptionsManager_OnEventRemoved(object sender, string eventName)
-        {
-            if (!_persistentConnection.IsConnected)
-            {
-                _persistentConnection.TryConnect();
-            }
-            using (var channel = _persistentConnection.CreateModel())
-            {
-                channel.QueueUnbind(queue: _queueName, exchange: BROKER_NAME, routingKey: eventName);
-                if (_subscriptionsManager.IsEmpty)
-                {
-                    _queueName = string.Empty;
-                    _consumerChannel.Close();
-                }
-            }
-        }
-
-        private IModel CreateConsumerChannel()
-        {
-            if (!_persistentConnection.IsConnected)
-            {
-                _persistentConnection.TryConnect();
-            }
-            _logger.LogTrace("Creating RabbitMQ consumer channel");
-
-            var channel = _persistentConnection.CreateModel();
+            _logger.LogTrace("Declaring RabbitMQ exchange to publish event: {EventId}", @event.Id);
 
             channel.ExchangeDeclare(exchange: BROKER_NAME, type: "direct");
 
-            channel.QueueDeclare(queue: _queueName, durable: true,
-                exclusive: false, autoDelete: false, arguments: null);
+            var body = JsonSerializer.SerializeToUtf8Bytes(@event, @event.GetType(), new JsonSerializerOptions { WriteIndented = true });
 
-            channel.CallbackException += (sender, ea) =>
+            policy.Execute(() =>
             {
-                _logger.LogWarning(ea.Exception, "Recreating RabbitMQ consumer channel");
-                _consumerChannel.Dispose();
-                _consumerChannel = CreateConsumerChannel();
-                StartBasicConsume();
-            };
-            return channel;
+                var properties = channel.CreateBasicProperties();
+                properties.DeliveryMode = 2;//persistent
+
+                    _logger.LogTrace("Publishing event to RabbitMQ: {EventId}", @event.Id);
+
+                channel.BasicPublish(
+                    exchange: BROKER_NAME,
+                    routingKey: eventName,
+                    mandatory: true,
+                    basicProperties: properties,
+                    body: body);
+            });
         }
+    }
+    public void Subscribe<T, TH>()
+        where T : IntegrationEvent
+        where TH : IIntegrationEventHandler<T>
+    {
+        var eventName = _subscriptionsManager.GetEventKey<T>();
+        DoInternalSubscription(eventName);
 
-        private void StartBasicConsume()
+        _logger.LogInformation("Subscribing to event {EventName} with {EventHandler}", eventName, typeof(TH));
+        _subscriptionsManager.AddSubsciption<T, TH>();
+        StartBasicConsume();
+    }
+    public void SubscribeDynamic<TH>(string eventName) where TH : IDynamicIntegrationEventHandler
+    {
+        _logger.LogInformation("Subscribing to dynamic event {EventName} with {EventHandler}", eventName, typeof(TH));
+        DoInternalSubscription(eventName);
+        _subscriptionsManager.AddDynamicSubScription<TH>(eventName);
+        StartBasicConsume();
+    }
+
+    public void Unsubscribe<T, TH>()
+        where T : IntegrationEvent
+        where TH : IIntegrationEventHandler<T>
+    {
+        var eventName = _subscriptionsManager.GetEventKey<T>();
+        _logger.LogInformation("Unsubscribing from event {EventName}", eventName);
+        _subscriptionsManager.RemoveSubscription<T, TH>();
+    }
+
+    public void UnsubscribeDynamic<TH>(string eventName) where TH : IDynamicIntegrationEventHandler
+    {
+        _logger.LogInformation("Unsubscribing from dynamic event {EventName}", eventName);
+        _subscriptionsManager.RemoveDynamicSubscription<TH>(eventName);
+    }
+    public void Dispose()
+    {
+        if (_consumerChannel != null)
         {
-            _logger.LogTrace("Starting RabbitMQ basic consume.");
-            if (_consumerChannel != null)
+            _consumerChannel.Dispose();
+        }
+        _subscriptionsManager.Clear();
+    }
+
+    private void DoInternalSubscription(string eventName)
+    {
+        var containsKey = _subscriptionsManager.HasSubscriptionsForEvent(eventName);
+        if (!containsKey)
+        {
+            if (!_persistentConnection.IsConnected)
             {
-                var consumer = new AsyncEventingBasicConsumer(_consumerChannel);
-                consumer.Received += Consumer_Received;
-                _consumerChannel.BasicConsume(queue: _queueName, autoAck: false, consumer: consumer);
+                _persistentConnection.TryConnect();
             }
-            else
+            _consumerChannel.QueueBind(queue: _queueName, exchange: BROKER_NAME, routingKey: eventName);
+        }
+    }
+
+    private void SubscriptionsManager_OnEventRemoved(object sender, string eventName)
+    {
+        if (!_persistentConnection.IsConnected)
+        {
+            _persistentConnection.TryConnect();
+        }
+        using (var channel = _persistentConnection.CreateModel())
+        {
+            channel.QueueUnbind(queue: _queueName, exchange: BROKER_NAME, routingKey: eventName);
+            if (_subscriptionsManager.IsEmpty)
             {
-                _logger.LogError("StartBasicConsume can't call on _consumerChannel == null");
+                _queueName = string.Empty;
+                _consumerChannel.Close();
             }
         }
+    }
 
-        private async Task Consumer_Received(object sender, BasicDeliverEventArgs @event)
+    private IModel CreateConsumerChannel()
+    {
+        if (!_persistentConnection.IsConnected)
         {
-            var eventName = @event.RoutingKey;
-            var message = Encoding.UTF8.GetString(@event.Body.Span);
-            try
+            _persistentConnection.TryConnect();
+        }
+        _logger.LogTrace("Creating RabbitMQ consumer channel");
+
+        var channel = _persistentConnection.CreateModel();
+
+        channel.ExchangeDeclare(exchange: BROKER_NAME, type: "direct");
+
+        channel.QueueDeclare(queue: _queueName, durable: true,
+            exclusive: false, autoDelete: false, arguments: null);
+
+        channel.CallbackException += (sender, ea) =>
+        {
+            _logger.LogWarning(ea.Exception, "Recreating RabbitMQ consumer channel");
+            _consumerChannel.Dispose();
+            _consumerChannel = CreateConsumerChannel();
+            StartBasicConsume();
+        };
+        return channel;
+    }
+
+    private void StartBasicConsume()
+    {
+        _logger.LogTrace("Starting RabbitMQ basic consume.");
+        if (_consumerChannel != null)
+        {
+            var consumer = new AsyncEventingBasicConsumer(_consumerChannel);
+            consumer.Received += Consumer_Received;
+            _consumerChannel.BasicConsume(queue: _queueName, autoAck: false, consumer: consumer);
+        }
+        else
+        {
+            _logger.LogError("StartBasicConsume can't call on _consumerChannel == null");
+        }
+    }
+
+    private async Task Consumer_Received(object sender, BasicDeliverEventArgs @event)
+    {
+        var eventName = @event.RoutingKey;
+        var message = Encoding.UTF8.GetString(@event.Body.Span);
+        try
+        {
+            if (message.ToLowerInvariant().Contains("throw-fake-exception"))
             {
-                if (message.ToLowerInvariant().Contains("throw-fake-exception"))
+                throw new InvalidOperationException($"Fake exception requested: \"{message}\"");
+            }
+            await ProcessEvent(eventName, message);
+        }
+        catch (Exception ex)
+        {
+            _logger.LogWarning(ex, "----- ERROR Processing message \"{Message}\"", message);
+        }
+    }
+
+    private async Task ProcessEvent(string eventName, string message)
+    {
+        _logger.LogTrace("Processing RabbitMQ event:{EventName}", eventName);
+
+        if (_subscriptionsManager.HasSubscriptionsForEvent(eventName))
+        {
+            using (var scope = _autofac.BeginLifetimeScope(AUTOFAC_SCOPE_NAME))
+            {
+                var subscriptions = _subscriptionsManager.GetHandlersForEvent(eventName);
+                foreach (var subscription in subscriptions)
                 {
-                    throw new InvalidOperationException($"Fake exception requested: \"{message}\"");
-                }
-                await ProcessEvent(eventName, message);
-            }
-            catch (Exception ex)
-            {
-                _logger.LogWarning(ex, "----- ERROR Processing message \"{Message}\"", message);
-            }
-        }
-
-        private async Task ProcessEvent(string eventName, string message)
-        {
-            _logger.LogTrace("Processing RabbitMQ event:{EventName}", eventName);
-
-            if (_subscriptionsManager.HasSubscriptionsForEvent(eventName))
-            {
-                using (var scope = _autofac.BeginLifetimeScope(AUTOFAC_SCOPE_NAME))
-                {
-                    var subscriptions = _subscriptionsManager.GetHandlersForEvent(eventName);
-                    foreach (var subscription in subscriptions)
+                    if (subscription.IsDynamic)
                     {
-                        if (subscription.IsDynamic)
+                        var handler = scope.ResolveOptional(subscription.HandlerType) as IDynamicIntegrationEventHandler;
+                        if (handler == null)
                         {
-                            var handler = scope.ResolveOptional(subscription.HandlerType) as IDynamicIntegrationEventHandler;
-                            if (handler == null)
-                            {
-                                continue;
-                            }
-                            using dynamic eventData = JsonDocument.Parse(message);
-                            await Task.Yield();
-                            await handler.Handle(eventData);
+                            continue;
                         }
-                        else
+                        using dynamic eventData = JsonDocument.Parse(message);
+                        await Task.Yield();
+                        await handler.Handle(eventData);
+                    }
+                    else
+                    {
+                        var handler = scope.ResolveOptional(subscription.HandlerType);
+                        if (handler == null)
                         {
-                            var handler = scope.ResolveOptional(subscription.HandlerType);
-                            if (handler == null)
-                            {
-                                continue;
-                            }
-                            var eventType = _subscriptionsManager.GetEventTypeByName(eventName);
-                            var integrationEvent = JsonSerializer.Deserialize(message, eventType, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
-                            var concreteType = typeof(IIntegrationEventHandler<>).MakeGenericType(eventType);
+                            continue;
+                        }
+                        var eventType = _subscriptionsManager.GetEventTypeByName(eventName);
+                        var integrationEvent = JsonSerializer.Deserialize(message, eventType, new JsonSerializerOptions { PropertyNameCaseInsensitive = true });
+                        var concreteType = typeof(IIntegrationEventHandler<>).MakeGenericType(eventType);
 
-                            await Task.Yield();
-                            await (Task)concreteType.GetMethod("Handle").Invoke(handler, new object[] { integrationEvent });
-                        }
+                        await Task.Yield();
+                        await (Task)concreteType.GetMethod("Handle").Invoke(handler, new object[] { integrationEvent });
                     }
                 }
             }
-            else
-            {
-                _logger.LogWarning("No subscription for RabbitMQ event: {EventName}", eventName);
-            }
+        }
+        else
+        {
+            _logger.LogWarning("No subscription for RabbitMQ event: {EventName}", eventName);
         }
     }
 }
