@@ -1,52 +1,26 @@
-﻿using Microsoft.Azure.ServiceBus;
-using System;
-
-namespace EventBusServiceBus
+﻿namespace EventBusServiceBus;
+public class DefaultServiceBusPersisterConnection : IServiceBusPersisterConnection
 {
-    public class DefaultServiceBusPersisterConnection : IServiceBusPersisterConnection
+    private readonly ServiceBusConnectionStringBuilder _serviceBusConnectionStringBuilder;
+    private readonly string _subscriptionClientName;
+    private SubscriptionClient _subscriptionClient;
+    private ITopicClient _topicClient;
+
+    bool _disposed;
+
+    public ServiceBusConnectionStringBuilder ServiceBusConnectionStringBuilder => _serviceBusConnectionStringBuilder;
+
+    public DefaultServiceBusPersisterConnection(ServiceBusConnectionStringBuilder serviceBusConnectionStringBuilder, string subscriptionClientName)
     {
-        private readonly ServiceBusConnectionStringBuilder _serviceBusConnectionStringBuilder;
-        private readonly string _subscriptionClientName;
-        private SubscriptionClient _subscriptionClient;
-        private ITopicClient _topicClient;
+        _serviceBusConnectionStringBuilder = serviceBusConnectionStringBuilder ?? throw new ArgumentNullException(nameof(serviceBusConnectionStringBuilder));
+        _subscriptionClientName = subscriptionClientName;
+        _subscriptionClient = new SubscriptionClient(_serviceBusConnectionStringBuilder, subscriptionClientName);
+        _topicClient = new TopicClient(_serviceBusConnectionStringBuilder, RetryPolicy.Default);
+    }
 
-        bool _disposed;
-
-        public ServiceBusConnectionStringBuilder ServiceBusConnectionStringBuilder => _serviceBusConnectionStringBuilder;
-
-        public DefaultServiceBusPersisterConnection(ServiceBusConnectionStringBuilder serviceBusConnectionStringBuilder, string subscriptionClientName)
-        {
-            _serviceBusConnectionStringBuilder = serviceBusConnectionStringBuilder ?? throw new ArgumentNullException(nameof(serviceBusConnectionStringBuilder));
-            _subscriptionClientName = subscriptionClientName;
-            _subscriptionClient = new SubscriptionClient(_serviceBusConnectionStringBuilder, subscriptionClientName);
-            _topicClient = new TopicClient(_serviceBusConnectionStringBuilder, RetryPolicy.Default);
-        }
-
-        public ITopicClient TopicClient
-        {
-            get
-            {
-                if (_topicClient.IsClosedOrClosing)
-                {
-                    _topicClient = new TopicClient(_serviceBusConnectionStringBuilder, RetryPolicy.Default);
-                }
-                return _topicClient;
-            }
-        }
-
-        public ISubscriptionClient SubscriptionClient
-        {
-            get
-            {
-                if (_subscriptionClient.IsClosedOrClosing)
-                {
-                    _subscriptionClient = new SubscriptionClient(_serviceBusConnectionStringBuilder, _subscriptionClientName);
-                }
-                return _subscriptionClient;
-            }
-        }
-
-        public ITopicClient CreateModel()
+    public ITopicClient TopicClient
+    {
+        get
         {
             if (_topicClient.IsClosedOrClosing)
             {
@@ -54,14 +28,35 @@ namespace EventBusServiceBus
             }
             return _topicClient;
         }
+    }
 
-        public void Dispose()
+    public ISubscriptionClient SubscriptionClient
+    {
+        get
         {
-            if (_disposed)
+            if (_subscriptionClient.IsClosedOrClosing)
             {
-                return;
+                _subscriptionClient = new SubscriptionClient(_serviceBusConnectionStringBuilder, _subscriptionClientName);
             }
-            _disposed = true;
+            return _subscriptionClient;
         }
+    }
+
+    public ITopicClient CreateModel()
+    {
+        if (_topicClient.IsClosedOrClosing)
+        {
+            _topicClient = new TopicClient(_serviceBusConnectionStringBuilder, RetryPolicy.Default);
+        }
+        return _topicClient;
+    }
+
+    public void Dispose()
+    {
+        if (_disposed)
+        {
+            return;
+        }
+        _disposed = true;
     }
 }
