@@ -1,193 +1,180 @@
-﻿using Autofac;
-using EventBus;
-using EventBus.Abstractions;
-using EventBus.Events;
-using Microsoft.Azure.ServiceBus;
-using Microsoft.Extensions.Logging;
-using System;
-using System.Text;
-using System.Text.Json;
-using System.Threading.Tasks;
-
-namespace EventBusServiceBus
+﻿namespace EventBusServiceBus;
+public class EventBusServiceBus : IEventBus
 {
-    public class EventBusServiceBus : IEventBus
+    private readonly IServiceBusPersisterConnection _serviceBusPersisterConnection;
+    private readonly ILogger<EventBusServiceBus> _logger;
+    private readonly IEventBusSubscriptionsManager _subscriptionsManager;
+    private readonly ILifetimeScope _autofac;
+    private readonly string AUTOFAC_SCOPE_NAME = "eshop_event_bus";
+    private const string INTEGRATION_EVENT_SUFFIX = "IntegrationEvent";
+
+    public EventBusServiceBus(IServiceBusPersisterConnection serviceBusPersisterConnection,
+        ILogger<EventBusServiceBus> logger, IEventBusSubscriptionsManager subscriptionsManager,
+        ILifetimeScope autofac)
     {
-        private readonly IServiceBusPersisterConnection _serviceBusPersisterConnection;
-        private readonly ILogger<EventBusServiceBus> _logger;
-        private readonly IEventBusSubscriptionsManager _subscriptionsManager;
-        private readonly ILifetimeScope _autofac;
-        private readonly string AUTOFAC_SCOPE_NAME = "eshop_event_bus";
-        private const string INTEGRATION_EVENT_SUFFIX = "IntegrationEvent";
+        _serviceBusPersisterConnection = serviceBusPersisterConnection;
+        _logger = logger ?? throw new ArgumentNullException(nameof(logger));
+        _subscriptionsManager = subscriptionsManager ?? new InMemoryEventBusSubscriptionsManager();
+        _autofac = autofac;
 
-        public EventBusServiceBus(IServiceBusPersisterConnection serviceBusPersisterConnection,
-            ILogger<EventBusServiceBus> logger, IEventBusSubscriptionsManager subscriptionsManager,
-            ILifetimeScope autofac)
+        RemoveDefaultRule();
+        RegisterSubscriptionClientMessageHandler();
+    }
+    public void Publish(IntegrationEvent @event)
+    {
+        var eventName = @event.GetType().Name.Replace(INTEGRATION_EVENT_SUFFIX, string.Empty);
+        var jsonMessage = JsonSerializer.Serialize(@event, @event.GetType());
+        var body = Encoding.UTF8.GetBytes(jsonMessage);
+
+        var message = new Message
         {
-            _serviceBusPersisterConnection = serviceBusPersisterConnection;
-            _logger = logger ?? throw new ArgumentNullException(nameof(logger));
-            _subscriptionsManager = subscriptionsManager ?? new InMemoryEventBusSubscriptionsManager();
-            _autofac = autofac;
+            MessageId = Guid.NewGuid().ToString(),
+            Body = body,
+            Label = eventName
+        };
 
-            RemoveDefaultRule();
-            RegisterSubscriptionClientMessageHandler();
-        }
-        public void Publish(IntegrationEvent @event)
+        _serviceBusPersisterConnection.TopicClient.SendAsync(message)
+            .GetAwaiter().GetResult();
+    }
+
+    public void Subscribe<T, TH>()
+        where T : IntegrationEvent
+        where TH : IIntegrationEventHandler<T>
+    {
+        var eventName = typeof(T).Name.Replace(INTEGRATION_EVENT_SUFFIX, string.Empty);
+        var containsKey = _subscriptionsManager.HasSubscriptionsForEvent<T>();
+        if (!containsKey)
         {
-            var eventName = @event.GetType().Name.Replace(INTEGRATION_EVENT_SUFFIX, string.Empty);
-            var jsonMessage = JsonSerializer.Serialize(@event, @event.GetType());
-            var body = Encoding.UTF8.GetBytes(jsonMessage);
-
-            var message = new Message
+            try
             {
-                MessageId = Guid.NewGuid().ToString(),
-                Body = body,
-                Label = eventName
-            };
+                _serviceBusPersisterConnection.SubscriptionClient.AddRuleAsync(new RuleDescription
+                {
+                    Filter = new CorrelationFilter { Label = eventName },
+                    Name = eventName
+                });
+            }
+            catch (ServiceBusException)
+            {
+                _logger.LogWarning("The message entity {EventName} already exists.", eventName);
+            }
+        }
+        _logger.LogInformation("Subscribing to event {EventName} with {EventHandler}", eventName, typeof(TH).Name);
+        _subscriptionsManager.AddSubsciption<T, TH>();
+    }
 
-            _serviceBusPersisterConnection.TopicClient.SendAsync(message)
+    public void SubscribeDynamic<TH>(string eventName)
+        where TH : IDynamicIntegrationEventHandler
+    {
+        _logger.LogInformation("Subscribing to dynamic event {EventName} with {EventHandler}", eventName, typeof(TH).Name);
+        _subscriptionsManager.AddDynamicSubScription<TH>(eventName);
+    }
+
+    public void Unsubscribe<T, TH>()
+        where T : IntegrationEvent
+        where TH : IIntegrationEventHandler<T>
+    {
+        var eventName = typeof(T).Name.Replace(INTEGRATION_EVENT_SUFFIX, string.Empty);
+        try
+        {
+            _serviceBusPersisterConnection
+                .SubscriptionClient
+                .RemoveRuleAsync(eventName)
                 .GetAwaiter().GetResult();
         }
-
-        public void Subscribe<T, TH>()
-            where T : IntegrationEvent
-            where TH : IIntegrationEventHandler<T>
+        catch (MessagingEntityNotFoundException)
         {
-            var eventName = typeof(T).Name.Replace(INTEGRATION_EVENT_SUFFIX, string.Empty);
-            var containsKey = _subscriptionsManager.HasSubscriptionsForEvent<T>();
-            if (!containsKey)
+            _logger.LogWarning("The messaging entity {EventName} could not be found.", eventName);
+        }
+        _logger.LogInformation("Unsubscribing from event {EventName}", eventName);
+        _subscriptionsManager.RemoveSubscription<T, TH>();
+    }
+
+    public void UnsubscribeDynamic<TH>(string eventName)
+        where TH : IDynamicIntegrationEventHandler
+    {
+        _logger.LogInformation("Unsubscribing from dynamic event {EventName}", eventName);
+        _subscriptionsManager.RemoveDynamicSubscription<TH>(eventName);
+    }
+
+    public void Dispose()
+    {
+        _subscriptionsManager.Clear();
+    }
+
+    private void RegisterSubscriptionClientMessageHandler()
+    {
+        _serviceBusPersisterConnection.SubscriptionClient.RegisterMessageHandler(
+            async (message, token) =>
             {
-                try
+                var eventName = $"{message.Label}{INTEGRATION_EVENT_SUFFIX}";
+                var messageData = Encoding.UTF8.GetString(message.Body);
+                if (await ProcessEvent(eventName, messageData))
                 {
-                    _serviceBusPersisterConnection.SubscriptionClient.AddRuleAsync(new RuleDescription
-                    {
-                        Filter = new CorrelationFilter { Label = eventName },
-                        Name = eventName
-                    });
+                    await _serviceBusPersisterConnection.SubscriptionClient.CompleteAsync(message.SystemProperties.LockToken);
                 }
-                catch (ServiceBusException)
-                {
-                    _logger.LogWarning("The message entity {EventName} already exists.", eventName);
-                }
-            }
-            _logger.LogInformation("Subscribing to event {EventName} with {EventHandler}", eventName, typeof(TH).Name);
-            _subscriptionsManager.AddSubsciption<T, TH>();
-        }
+            },
+            new MessageHandlerOptions(ExceptionReceivedHandler) { MaxConcurrentCalls = 10, AutoComplete = false });
+    }
 
-        public void SubscribeDynamic<TH>(string eventName)
-            where TH : IDynamicIntegrationEventHandler
-        {
-            _logger.LogInformation("Subscribing to dynamic event {EventName} with {EventHandler}", eventName, typeof(TH).Name);
-            _subscriptionsManager.AddDynamicSubScription<TH>(eventName);
-        }
+    private Task ExceptionReceivedHandler(ExceptionReceivedEventArgs exceptionReceivedEventArgs)
+    {
+        var ex = exceptionReceivedEventArgs.Exception;
+        var context = exceptionReceivedEventArgs.ExceptionReceivedContext;
+        _logger.LogError(ex, "ERROR handling message: {ExceptionMessage} - Context :{@ExceptionContext}", ex.Message, context);
+        return Task.CompletedTask;
+    }
 
-        public void Unsubscribe<T, TH>()
-            where T : IntegrationEvent
-            where TH : IIntegrationEventHandler<T>
+    private async Task<bool> ProcessEvent(string eventName, string message)
+    {
+        var processed = false;
+        if (_subscriptionsManager.HasSubscriptionsForEvent(eventName))
         {
-            var eventName = typeof(T).Name.Replace(INTEGRATION_EVENT_SUFFIX, string.Empty);
-            try
+            using (var scope = _autofac.BeginLifetimeScope(AUTOFAC_SCOPE_NAME))
             {
-                _serviceBusPersisterConnection
-                    .SubscriptionClient
-                    .RemoveRuleAsync(eventName)
-                    .GetAwaiter().GetResult();
-            }
-            catch (MessagingEntityNotFoundException)
-            {
-                _logger.LogWarning("The messaging entity {EventName} could not be found.", eventName);
-            }
-            _logger.LogInformation("Unsubscribing from event {EventName}", eventName);
-            _subscriptionsManager.RemoveSubscription<T, TH>();
-        }
-
-        public void UnsubscribeDynamic<TH>(string eventName)
-            where TH : IDynamicIntegrationEventHandler
-        {
-            _logger.LogInformation("Unsubscribing from dynamic event {EventName}", eventName);
-            _subscriptionsManager.RemoveDynamicSubscription<TH>(eventName);
-        }
-
-        public void Dispose()
-        {
-            _subscriptionsManager.Clear();
-        }
-
-        private void RegisterSubscriptionClientMessageHandler()
-        {
-            _serviceBusPersisterConnection.SubscriptionClient.RegisterMessageHandler(
-                async (message, token) =>
+                var subscriptions = _subscriptionsManager.GetHandlersForEvent(eventName);
+                foreach (var subscription in subscriptions)
                 {
-                    var eventName = $"{message.Label}{INTEGRATION_EVENT_SUFFIX}";
-                    var messageData = Encoding.UTF8.GetString(message.Body);
-                    if (await ProcessEvent(eventName, messageData))
+                    if (subscription.IsDynamic)
                     {
-                        await _serviceBusPersisterConnection.SubscriptionClient.CompleteAsync(message.SystemProperties.LockToken);
+                        var handler = scope.ResolveOptional(subscription.HandlerType) as IDynamicIntegrationEventHandler;
+                        if (handler == null)
+                        {
+                            continue;
+                        }
+
+                        using dynamic eventData = JsonDocument.Parse(message);
+                        await handler.Handle(eventData);
                     }
-                },
-                new MessageHandlerOptions(ExceptionReceivedHandler) { MaxConcurrentCalls = 10, AutoComplete = false });
-        }
-
-        private Task ExceptionReceivedHandler(ExceptionReceivedEventArgs exceptionReceivedEventArgs)
-        {
-            var ex = exceptionReceivedEventArgs.Exception;
-            var context = exceptionReceivedEventArgs.ExceptionReceivedContext;
-            _logger.LogError(ex, "ERROR handling message: {ExceptionMessage} - Context :{@ExceptionContext}", ex.Message, context);
-            return Task.CompletedTask;
-        }
-
-        private async Task<bool> ProcessEvent(string eventName, string message)
-        {
-            var processed = false;
-            if (_subscriptionsManager.HasSubscriptionsForEvent(eventName))
-            {
-                using (var scope = _autofac.BeginLifetimeScope(AUTOFAC_SCOPE_NAME))
-                {
-                    var subscriptions = _subscriptionsManager.GetHandlersForEvent(eventName);
-                    foreach (var subscription in subscriptions)
+                    else
                     {
-                        if (subscription.IsDynamic)
+                        var handler = scope.ResolveOptional(subscription.HandlerType);
+                        if (handler == null)
                         {
-                            var handler = scope.ResolveOptional(subscription.HandlerType) as IDynamicIntegrationEventHandler;
-                            if (handler == null)
-                            {
-                                continue;
-                            }
-
-                            using dynamic eventData = JsonDocument.Parse(message);
-                            await handler.Handle(eventData);
+                            continue;
                         }
-                        else
-                        {
-                            var handler = scope.ResolveOptional(subscription.HandlerType);
-                            if (handler == null)
-                            {
-                                continue;
-                            }
-                            var eventType = _subscriptionsManager.GetEventTypeByName(eventName);
-                            var integrationEvent = JsonSerializer.Deserialize(message, eventType);
-                            var concreteType = typeof(IIntegrationEventHandler<>).MakeGenericType(eventType);
-                            await (Task)concreteType.GetMethod("Handle").Invoke(handler, new object[] { integrationEvent });
-                        }
+                        var eventType = _subscriptionsManager.GetEventTypeByName(eventName);
+                        var integrationEvent = JsonSerializer.Deserialize(message, eventType);
+                        var concreteType = typeof(IIntegrationEventHandler<>).MakeGenericType(eventType);
+                        await (Task)concreteType.GetMethod("Handle").Invoke(handler, new object[] { integrationEvent });
                     }
                 }
-                processed = true;
             }
-            return processed;
+            processed = true;
         }
+        return processed;
+    }
 
-        private void RemoveDefaultRule()
+    private void RemoveDefaultRule()
+    {
+        try
         {
-            try
-            {
-                _serviceBusPersisterConnection.SubscriptionClient
-                    .RemoveRuleAsync(RuleDescription.DefaultRuleName)
-                    .GetAwaiter().GetResult();
-            }
-            catch (MessagingEntityNotFoundException)
-            {
-                _logger.LogWarning("The messaging entity {DefaultRuleName} could not be found.", RuleDescription.DefaultRuleName);
-            }
+            _serviceBusPersisterConnection.SubscriptionClient
+                .RemoveRuleAsync(RuleDescription.DefaultRuleName)
+                .GetAwaiter().GetResult();
+        }
+        catch (MessagingEntityNotFoundException)
+        {
+            _logger.LogWarning("The messaging entity {DefaultRuleName} could not be found.", RuleDescription.DefaultRuleName);
         }
     }
 }
