@@ -1,40 +1,30 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace IntegrationEventLogEF
+﻿namespace IntegrationEventLogEF;
+public class IntegrationEventLogContext : DbContext
 {
-    public class IntegrationEventLogContext : DbContext
+    public IntegrationEventLogContext(DbContextOptions<IntegrationEventLogContext> options)
+        : base(options)
     {
-        public IntegrationEventLogContext(DbContextOptions<IntegrationEventLogContext> options)
-            : base(options)
-        {
 
-        }
+    }
 
-        public DbSet<IntegrationEventLogEntry> IntegrationEventLogs { get; set; }
+    public DbSet<IntegrationEventLogEntry> IntegrationEventLogs { get; set; }
 
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.Entity<IntegrationEventLogEntry>(ConfigureIntegrationEventLogEntry);
-        }
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.Entity<IntegrationEventLogEntry>(ConfigureIntegrationEventLogEntry);
+    }
 
-        void ConfigureIntegrationEventLogEntry(EntityTypeBuilder<IntegrationEventLogEntry> builder)
-        {
-            builder.ToTable("IntegrationEventLog");
+    void ConfigureIntegrationEventLogEntry(EntityTypeBuilder<IntegrationEventLogEntry> builder)
+    {
+        builder.ToTable("IntegrationEventLog");
 
-            builder.HasKey(e => e.EventId);
+        builder.HasKey(e => e.EventId);
 
-            builder.Property(e => e.EventId).IsRequired();
-            builder.Property(e => e.Content).IsRequired();
-            builder.Property(e => e.CreationTime).IsRequired();
-            builder.Property(e => e.State).IsRequired();
-            builder.Property(e => e.TimesSent).IsRequired();
-            builder.Property(e => e.EventTypeName).IsRequired();
-        }
+        builder.Property(e => e.EventId).IsRequired();
+        builder.Property(e => e.Content).IsRequired();
+        builder.Property(e => e.CreationTime).IsRequired();
+        builder.Property(e => e.State).IsRequired();
+        builder.Property(e => e.TimesSent).IsRequired();
+        builder.Property(e => e.EventTypeName).IsRequired();
     }
 }
