@@ -1,45 +1,37 @@
-﻿using EventBus.Events;
-using System;
-using System.ComponentModel.DataAnnotations.Schema;
-using System.Linq;
-using System.Text.Json;
-
-namespace IntegrationEventLogEF
+﻿namespace IntegrationEventLogEF;
+public class IntegrationEventLogEntry
 {
-    public class IntegrationEventLogEntry
+    private IntegrationEventLogEntry() { }
+
+    public IntegrationEventLogEntry(IntegrationEvent @event, Guid transactionId)
     {
-        private IntegrationEventLogEntry() { }
-
-        public IntegrationEventLogEntry(IntegrationEvent @event, Guid transactionId)
+        EventId = @event.Id;
+        CreationTime = @event.CreationDate;
+        EventTypeName = @event.GetType().FullName;
+        Content = JsonSerializer.Serialize(@event, @event.GetType(), new JsonSerializerOptions
         {
-            EventId = @event.Id;
-            CreationTime = @event.CreationDate;
-            EventTypeName = @event.GetType().FullName;
-            Content = JsonSerializer.Serialize(@event, @event.GetType(), new JsonSerializerOptions
-            {
-                WriteIndented = true
-            });
-            State = EventStateEnum.NotPublished;
-            TimesSent = 0;
-            TransactionId = TransactionId.ToString();
-        }
+            WriteIndented = true
+        });
+        State = EventStateEnum.NotPublished;
+        TimesSent = 0;
+        TransactionId = TransactionId.ToString();
+    }
 
-        public Guid EventId { get; private set; }
-        public string EventTypeName { get; private set; }
-        [NotMapped]
-        public string EventTypeShortName => EventTypeName.Split(',')?.Last();
-        [NotMapped]
-        public IntegrationEvent IntegrationEvent { get; private set; }
-        public EventStateEnum State { get; set; }
-        public int TimesSent { get; set; }
-        public DateTime CreationTime { get; private set; }
-        public string Content { get; private set; }
-        public string TransactionId { get; private set; }
+    public Guid EventId { get; private set; }
+    public string EventTypeName { get; private set; }
+    [NotMapped]
+    public string EventTypeShortName => EventTypeName.Split(',')?.Last();
+    [NotMapped]
+    public IntegrationEvent IntegrationEvent { get; private set; }
+    public EventStateEnum State { get; set; }
+    public int TimesSent { get; set; }
+    public DateTime CreationTime { get; private set; }
+    public string Content { get; private set; }
+    public string TransactionId { get; private set; }
 
-        public IntegrationEventLogEntry DeserializeJsonContent(Type type)
-        {
-            IntegrationEvent = JsonSerializer.Deserialize(Content, type, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) as IntegrationEvent;
-            return this;
-        }
+    public IntegrationEventLogEntry DeserializeJsonContent(Type type)
+    {
+        IntegrationEvent = JsonSerializer.Deserialize(Content, type, new JsonSerializerOptions { PropertyNameCaseInsensitive = true }) as IntegrationEvent;
+        return this;
     }
 }
