@@ -1,17 +1,9 @@
-﻿using EventBus.Events;
-using Microsoft.EntityFrameworkCore.Storage;
-using System;
-using System.Collections.Generic;
-using System.Threading.Tasks;
-
-namespace IntegrationEventLogEF.Services
+﻿namespace IntegrationEventLogEF.Services;
+public interface IIntegrationEventLogService
 {
-    public interface IIntegrationEventLogService
-    {
-        Task<IEnumerable<IntegrationEventLogEntry>> RetrieveEventLogsPendingToPublishAsync(Guid transactionId);
-        Task SaveEventAsync(IntegrationEvent @event, IDbContextTransaction transaction);
-        Task MarkEventAsPublishedAsync(Guid eventId);
-        Task MarkEventAsInProgressAsync(Guid eventId);
-        Task MarkEventAsFailedAsync(Guid eventId);
-    }
+    Task<IEnumerable<IntegrationEventLogEntry>> RetrieveEventLogsPendingToPublishAsync(Guid transactionId);
+    Task SaveEventAsync(IntegrationEvent @event, IDbContextTransaction transaction);
+    Task MarkEventAsPublishedAsync(Guid eventId);
+    Task MarkEventAsInProgressAsync(Guid eventId);
+    Task MarkEventAsFailedAsync(Guid eventId);
 }
