@@ -1,109 +1,97 @@
-﻿using EventBus.Events;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Storage;
-using System;
-using System.Collections.Generic;
-using System.Data.Common;
-using System.Linq;
-using System.Reflection;
-using System.Threading.Tasks;
-
-namespace IntegrationEventLogEF.Services
+﻿namespace IntegrationEventLogEF.Services;
+public class IntegrationEventLogService : IIntegrationEventLogService
 {
-    public class IntegrationEventLogService : IIntegrationEventLogService
+    private readonly IntegrationEventLogContext _integrationEventLogContext;
+    private readonly DbConnection _dbConnection;
+    private readonly List<Type> _eventTypes;
+    private volatile bool disposedValue;
+
+    public IntegrationEventLogService(DbConnection dbConnection)
     {
-        private readonly IntegrationEventLogContext _integrationEventLogContext;
-        private readonly DbConnection _dbConnection;
-        private readonly List<Type> _eventTypes;
-        private volatile bool disposedValue;
+        _dbConnection = dbConnection ?? throw new ArgumentNullException(nameof(dbConnection));
+        _integrationEventLogContext = new IntegrationEventLogContext(
+            new DbContextOptionsBuilder<IntegrationEventLogContext>()
+            .UseSqlServer(_dbConnection).Options);
 
-        public IntegrationEventLogService(DbConnection dbConnection)
+        _eventTypes = Assembly.Load(Assembly.GetEntryAssembly().FullName)
+            .GetTypes()
+            .Where(t => t.Name.EndsWith(nameof(IntegrationEvent)))
+            .ToList();
+    }
+    public async Task<IEnumerable<IntegrationEventLogEntry>> RetrieveEventLogsPendingToPublishAsync(Guid transactionId)
+    {
+        var tid = transactionId.ToString();
+
+        var result = await _integrationEventLogContext.IntegrationEventLogs
+            .Where(e => e.TransactionId == tid && e.State == EventStateEnum.NotPublished).ToListAsync();
+        if (result != null && result.Any())
         {
-            _dbConnection = dbConnection ?? throw new ArgumentNullException(nameof(dbConnection));
-            _integrationEventLogContext = new IntegrationEventLogContext(
-                new DbContextOptionsBuilder<IntegrationEventLogContext>()
-                .UseSqlServer(_dbConnection).Options);
-
-            _eventTypes = Assembly.Load(Assembly.GetEntryAssembly().FullName)
-                .GetTypes()
-                .Where(t => t.Name.EndsWith(nameof(IntegrationEvent)))
-                .ToList();
+            return result.OrderBy(o => o.CreationTime)
+                .Select(e => e.DeserializeJsonContent(_eventTypes.Find(t => t.Name == e.EventTypeShortName)));
         }
-        public async Task<IEnumerable<IntegrationEventLogEntry>> RetrieveEventLogsPendingToPublishAsync(Guid transactionId)
-        {
-            var tid = transactionId.ToString();
 
-            var result = await _integrationEventLogContext.IntegrationEventLogs
-                .Where(e => e.TransactionId == tid && e.State == EventStateEnum.NotPublished).ToListAsync();
-            if (result != null && result.Any())
+        return new List<IntegrationEventLogEntry>();
+    }
+
+    public Task SaveEventAsync(IntegrationEvent @event, IDbContextTransaction transaction)
+    {
+        if (transaction == null)
+        {
+            throw new ArgumentNullException(nameof(transaction));
+        }
+        var eventLogEntry = new IntegrationEventLogEntry(@event, transaction.TransactionId);
+
+        _integrationEventLogContext.Database.UseTransaction(transaction.GetDbTransaction());
+        _integrationEventLogContext.IntegrationEventLogs.Add(eventLogEntry);
+
+        return _integrationEventLogContext.SaveChangesAsync();
+    }
+
+    public Task MarkEventAsFailedAsync(Guid eventId)
+    {
+        return UpdateEventStatus(eventId, EventStateEnum.PublishedFaild);
+    }
+
+    public Task MarkEventAsInProgressAsync(Guid eventId)
+    {
+        return UpdateEventStatus(eventId, EventStateEnum.InProgress);
+    }
+
+    public Task MarkEventAsPublishedAsync(Guid eventId)
+    {
+        return UpdateEventStatus(eventId, EventStateEnum.Published);
+    }
+
+    private Task UpdateEventStatus(Guid eventId, EventStateEnum status)
+    {
+        var eventLogEntry = _integrationEventLogContext.IntegrationEventLogs.Single(ie => ie.EventId == eventId);
+        eventLogEntry.State = status;
+
+        if (status == EventStateEnum.InProgress)
+        {
+            eventLogEntry.TimesSent++;
+        }
+
+        _integrationEventLogContext.IntegrationEventLogs.Update(eventLogEntry);
+
+        return _integrationEventLogContext.SaveChangesAsync();
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (!disposedValue)
+        {
+            if (disposing)
             {
-                return result.OrderBy(o => o.CreationTime)
-                    .Select(e => e.DeserializeJsonContent(_eventTypes.Find(t => t.Name == e.EventTypeShortName)));
+                _integrationEventLogContext?.Dispose();
             }
-
-            return new List<IntegrationEventLogEntry>();
+            disposedValue = true;
         }
+    }
 
-        public Task SaveEventAsync(IntegrationEvent @event, IDbContextTransaction transaction)
-        {
-            if (transaction == null)
-            {
-                throw new ArgumentNullException(nameof(transaction));
-            }
-            var eventLogEntry = new IntegrationEventLogEntry(@event, transaction.TransactionId);
-
-            _integrationEventLogContext.Database.UseTransaction(transaction.GetDbTransaction());
-            _integrationEventLogContext.IntegrationEventLogs.Add(eventLogEntry);
-
-            return _integrationEventLogContext.SaveChangesAsync();
-        }
-
-        public Task MarkEventAsFailedAsync(Guid eventId)
-        {
-            return UpdateEventStatus(eventId, EventStateEnum.PublishedFaild);
-        }
-
-        public Task MarkEventAsInProgressAsync(Guid eventId)
-        {
-            return UpdateEventStatus(eventId, EventStateEnum.InProgress);
-        }
-
-        public Task MarkEventAsPublishedAsync(Guid eventId)
-        {
-            return UpdateEventStatus(eventId, EventStateEnum.Published);
-        }
-
-        private Task UpdateEventStatus(Guid eventId, EventStateEnum status)
-        {
-            var eventLogEntry = _integrationEventLogContext.IntegrationEventLogs.Single(ie => ie.EventId == eventId);
-            eventLogEntry.State = status;
-
-            if (status == EventStateEnum.InProgress)
-            {
-                eventLogEntry.TimesSent++;
-            }
-
-            _integrationEventLogContext.IntegrationEventLogs.Update(eventLogEntry);
-
-            return _integrationEventLogContext.SaveChangesAsync();
-        }
-
-        protected virtual void Dispose(bool disposing)
-        {
-            if (!disposedValue)
-            {
-                if (disposing)
-                {
-                    _integrationEventLogContext?.Dispose();
-                }
-                disposedValue = true;
-            }
-        }
-
-        public void Dispose()
-        {
-            Dispose(disposing: true);
-            GC.SuppressFinalize(this);
-        }
+    public void Dispose()
+    {
+        Dispose(disposing: true);
+        GC.SuppressFinalize(this);
     }
 }
