@@ -1,10 +1,8 @@
-﻿namespace Catalog.API
+﻿namespace Catalog.API;
+public class CatalogSettings
 {
-    public class CatalogSettings
-    {
-        public string PicBaseUrl { get; set; }
-        public string EventBusConnection { get; set; }
-        public bool UseCustomizationData { get; set; }
-        public bool AzureStorageEnabled { get; set; }
-    }
+    public string PicBaseUrl { get; set; }
+    public string EventBusConnection { get; set; }
+    public bool UseCustomizationData { get; set; }
+    public bool AzureStorageEnabled { get; set; }
 }
