@@ -1,12 +1,8 @@
-﻿using Microsoft.AspNetCore.Mvc;
-
-namespace Catalog.API.Controllers
+﻿namespace Catalog.API.Controllers;
+public class HomeController : Controller
 {
-    public class HomeController : Controller
+    public IActionResult Index()
     {
-        public IActionResult Index()
-        {
-            return new RedirectResult("~/swagger");
-        }
+        return new RedirectResult("~/swagger");
     }
 }
