@@ -1,37 +1,30 @@
-﻿using Microsoft.EntityFrameworkCore;
-using Catalog.API.Model;
-using Catalog.API.Infrastructure.EntityConfigurations;
-using Microsoft.EntityFrameworkCore.Design;
-
-namespace Catalog.API.Infrastructure
+﻿namespace Catalog.API.Infrastructure;
+public class CatalogContext : DbContext
 {
-    public class CatalogContext : DbContext
+    public CatalogContext(DbContextOptions<CatalogContext> options) : base(options)
     {
-        public CatalogContext(DbContextOptions<CatalogContext> options) : base(options)
+
+    }
+
+    public DbSet<CatalogItem> CatalogItems { get; set; }
+    public DbSet<CatalogBrand> CatalogBrands { get; set; }
+    public DbSet<CatalogType> CatalogTypes { get; set; }
+
+    protected override void OnModelCreating(ModelBuilder modelBuilder)
+    {
+        modelBuilder.ApplyConfiguration(new CatalogBrandEntityTypeConfiguration());
+        modelBuilder.ApplyConfiguration(new CatalogTypeEntityTypeConfiguration());
+        modelBuilder.ApplyConfiguration(new CatalogItemEntityTypeConfiguration());
+    }
+
+    public class CatalogContextDesignFactory : IDesignTimeDbContextFactory<CatalogContext>
+    {
+        public CatalogContext CreateDbContext(string[] args)
         {
+            var optionsBuilder = new DbContextOptionsBuilder<CatalogContext>()
+                .UseSqlServer("Server=.;Initial Catalog=Microsoft.eShopWithoutContainers.Services.CatalogDb;Integrated Security=true");
 
-        }
-
-        public DbSet<CatalogItem> CatalogItems { get; set; }
-        public DbSet<CatalogBrand> CatalogBrands { get; set; }
-        public DbSet<CatalogType> CatalogTypes { get; set; }
-
-        protected override void OnModelCreating(ModelBuilder modelBuilder)
-        {
-            modelBuilder.ApplyConfiguration(new CatalogBrandEntityTypeConfiguration());
-            modelBuilder.ApplyConfiguration(new CatalogTypeEntityTypeConfiguration());
-            modelBuilder.ApplyConfiguration(new CatalogItemEntityTypeConfiguration());
-        }
-
-        public class CatalogContextDesignFactory : IDesignTimeDbContextFactory<CatalogContext>
-        {
-            public CatalogContext CreateDbContext(string[] args)
-            {
-                var optionsBuilder = new DbContextOptionsBuilder<CatalogContext>()
-                    .UseSqlServer("Server=.;Initial Catalog=Microsoft.eShopWithoutContainers.Services.CatalogDb;Integrated Security=true");
-
-                return new CatalogContext(optionsBuilder.Options);
-            }
+            return new CatalogContext(optionsBuilder.Options);
         }
     }
 }
