@@ -1,38 +1,32 @@
-﻿using Catalog.API.Model;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Catalog.API.Infrastructure.EntityConfigurations
+﻿namespace Catalog.API.Infrastructure.EntityConfigurations;
+public class CatalogItemEntityTypeConfiguration : IEntityTypeConfiguration<CatalogItem>
 {
-    public class CatalogItemEntityTypeConfiguration : IEntityTypeConfiguration<CatalogItem>
+    public void Configure(EntityTypeBuilder<CatalogItem> builder)
     {
-        public void Configure(EntityTypeBuilder<CatalogItem> builder)
-        {
-            builder.ToTable("Catalog");
+        builder.ToTable("Catalog");
 
-            builder.Property(ci => ci.Id)
-                .UseHiLo("catalog_hilo")
-                .IsRequired();
+        builder.Property(ci => ci.Id)
+            .UseHiLo("catalog_hilo")
+            .IsRequired();
 
-            builder.Property(ci => ci.Name)
-                .IsRequired()
-                .HasMaxLength(50);
+        builder.Property(ci => ci.Name)
+            .IsRequired()
+            .HasMaxLength(50);
 
-            builder.Property(ci => ci.Price)
-                .IsRequired();
+        builder.Property(ci => ci.Price)
+            .IsRequired();
 
-            builder.Property(ci => ci.PictureFileName)
-                .IsRequired(false);
+        builder.Property(ci => ci.PictureFileName)
+            .IsRequired(false);
 
-            builder.Ignore(ci => ci.PictureUri);
+        builder.Ignore(ci => ci.PictureUri);
 
-            builder.HasOne(ci => ci.CatalogBrand)
-                .WithMany()
-                .HasForeignKey(ci => ci.CatalogBrandId);
+        builder.HasOne(ci => ci.CatalogBrand)
+            .WithMany()
+            .HasForeignKey(ci => ci.CatalogBrandId);
 
-            builder.HasOne(ci => ci.CatalogType)
-                .WithMany()
-                .HasForeignKey(ci => ci.CatalogTypeId);
-        }
+        builder.HasOne(ci => ci.CatalogType)
+            .WithMany()
+            .HasForeignKey(ci => ci.CatalogTypeId);
     }
 }
