@@ -1,24 +1,18 @@
-﻿using Catalog.API.Model;
-using Microsoft.EntityFrameworkCore;
-using Microsoft.EntityFrameworkCore.Metadata.Builders;
-
-namespace Catalog.API.Infrastructure.EntityConfigurations
+﻿namespace Catalog.API.Infrastructure.EntityConfigurations;
+public class CatalogTypeEntityTypeConfiguration : IEntityTypeConfiguration<CatalogType>
 {
-    public class CatalogTypeEntityTypeConfiguration : IEntityTypeConfiguration<CatalogType>
+    public void Configure(EntityTypeBuilder<CatalogType> builder)
     {
-        public void Configure(EntityTypeBuilder<CatalogType> builder)
-        {
-            builder.ToTable("CatalogType");
+        builder.ToTable("CatalogType");
 
-            builder.HasKey(ci => ci.Id);
+        builder.HasKey(ci => ci.Id);
 
-            builder.Property(ci => ci.Id)
-                .UseHiLo("catalog_type_hilo")
-                .IsRequired();
+        builder.Property(ci => ci.Id)
+            .UseHiLo("catalog_type_hilo")
+            .IsRequired();
 
-            builder.Property(cb => cb.Type)
-                .IsRequired()
-                .HasMaxLength(100);
-        }
+        builder.Property(cb => cb.Type)
+            .IsRequired()
+            .HasMaxLength(100);
     }
 }
