@@ -1,8 +1,6 @@
-﻿namespace Catalog.API.Model
+﻿namespace Catalog.API.Model;
+public class CatalogBrand
 {
-    public class CatalogBrand
-    {
-        public int Id { get; set; }
-        public string Brand { get; set; }
-    }
+    public int Id { get; set; }
+    public string Brand { get; set; }
 }
