@@ -1,8 +1,6 @@
-﻿namespace Catalog.API.Model
+﻿namespace Catalog.API.Model;
+public class CatalogType
 {
-    public class CatalogType
-    {
-        public int Id { get; set; }
-        public string Type { get; set; }
-    }
+    public int Id { get; set; }
+    public string Type { get; set; }
 }
