@@ -1,10 +1,6 @@
-using System;
-
-namespace Identity.API
+namespace Identity.API;
+public class AppSettings
 {
-    public class AppSettings
-    {
-        public string MvcClient { get; set; }
-        public bool UseCustomizationData { get; set; }
-    }
+    public string MvcClient { get; set; }
+    public bool UseCustomizationData { get; set; }
 }
