@@ -1,15 +1,10 @@
-using IdentityServer4;
-using IdentityServer4.Models;
-using System.Collections.Generic;
-
-namespace Identity.API
+namespace Identity.API;
+public class Config
 {
-    public class Config
+    // ApiResources define the apis in your system
+    public static IEnumerable<ApiResource> GetApis()
     {
-        // ApiResources define the apis in your system
-        public static IEnumerable<ApiResource> GetApis()
-        {
-            return new List<ApiResource>
+        return new List<ApiResource>
             {
                 new ApiResource("orders", "Orders Service"),
                 new ApiResource("basket", "Basket Service"),
@@ -18,23 +13,23 @@ namespace Identity.API
                 new ApiResource("orders.signalrhub", "Ordering Signalr Hub"),
                 new ApiResource("webhooks", "Webhooks registration Service"),
             };
-        }
+    }
 
-        // Identity resources are data like user ID, name, or email address of a user
-        // see: http://docs.identityserver.io/en/release/configuration/resources.html
-        public static IEnumerable<IdentityResource> GetResources()
-        {
-            return new List<IdentityResource>
+    // Identity resources are data like user ID, name, or email address of a user
+    // see: http://docs.identityserver.io/en/release/configuration/resources.html
+    public static IEnumerable<IdentityResource> GetResources()
+    {
+        return new List<IdentityResource>
             {
                 new IdentityResources.OpenId(),
                 new IdentityResources.Profile()
             };
-        }
+    }
 
-        // client want to access resources (aka scopes)
-        public static IEnumerable<Client> GetClients(Dictionary<string, string> clientsUrl)
-        {
-            return new List<Client>
+    // client want to access resources (aka scopes)
+    public static IEnumerable<Client> GetClients(Dictionary<string, string> clientsUrl)
+    {
+        return new List<Client>
             {
                 // JavaScript Client
                 new Client
@@ -264,6 +259,5 @@ namespace Identity.API
                     }
                 }
             };
-        }
     }
 }
