@@ -1,232 +1,216 @@
-﻿using Identity.API.Models;
-using IdentityServer4;
-using IdentityServer4.Events;
-using IdentityServer4.Extensions;
-using IdentityServer4.Models;
-using IdentityServer4.Services;
-using IdentityServer4.Validation;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Logging;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Identity.API.Controllers
+﻿namespace Identity.API.Controllers;
+//[SecurityHeaders]
+[Authorize]
+public class ConsentController : Controller
 {
-    //[SecurityHeaders]
-    [Authorize]
-    public class ConsentController : Controller
+    private readonly IIdentityServerInteractionService _interactionService;
+    private readonly IEventService _eventService;
+    private readonly ILogger<ConsentController> _logger;
+
+    public ConsentController(
+        IIdentityServerInteractionService interactionService,
+        IEventService eventService,
+        ILogger<ConsentController> logger)
     {
-        private readonly IIdentityServerInteractionService _interactionService;
-        private readonly IEventService _eventService;
-        private readonly ILogger<ConsentController> _logger;
+        _interactionService = interactionService;
+        _eventService = eventService;
+        _logger = logger;
+    }
 
-        public ConsentController(
-            IIdentityServerInteractionService interactionService,
-            IEventService eventService,
-            ILogger<ConsentController> logger)
+    [HttpGet]
+    public async Task<IActionResult> Index(string returnUrl)
+    {
+        var vm = await BuildViewModelAsync(returnUrl);
+        if (vm is not null)
         {
-            _interactionService = interactionService;
-            _eventService = eventService;
-            _logger = logger;
+            return View("Index", vm);
+        }
+        return View("Error");
+    }
+
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Index(ConsentInputModel model)
+    {
+        var result = await ProcessConsent(model);
+        if (result.IsRedirect)
+        {
+            var context = await _interactionService.GetAuthorizationContextAsync(model.ReturnUrl);
+
+            if (context?.IsNativeClient() == true)
+            {
+                return this.LoadingPage("Redirect", result.RedirectUri);
+            }
+
+            return Redirect(result.RedirectUri);
         }
 
-        [HttpGet]
-        public async Task<IActionResult> Index(string returnUrl)
+        if (result.HasValidationError)
         {
-            var vm = await BuildViewModelAsync(returnUrl);
-            if (vm is not null)
-            {
-                return View("Index", vm);
-            }
-            return View("Error");
+            ModelState.AddModelError(string.Empty, result.ValidationError);
         }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Index(ConsentInputModel model)
+        if (result.ShowView)
         {
-            var result = await ProcessConsent(model);
-            if (result.IsRedirect)
-            {
-                var context = await _interactionService.GetAuthorizationContextAsync(model.ReturnUrl);
-
-                if (context?.IsNativeClient() == true)
-                {
-                    return this.LoadingPage("Redirect", result.RedirectUri);
-                }
-
-                return Redirect(result.RedirectUri);
-            }
-
-            if (result.HasValidationError)
-            {
-                ModelState.AddModelError(string.Empty, result.ValidationError);
-            }
-
-            if (result.ShowView)
-            {
-                return View("Index", result.ViewModel);
-            }
-            return View("Error");
+            return View("Index", result.ViewModel);
         }
+        return View("Error");
+    }
 
 
-        private async Task<ProcessConsentResult> ProcessConsent(ConsentInputModel model)
+    private async Task<ProcessConsentResult> ProcessConsent(ConsentInputModel model)
+    {
+        var result = new ProcessConsentResult();
+
+        var request = await _interactionService.GetAuthorizationContextAsync(model.ReturnUrl);
+        if (request is null)
         {
-            var result = new ProcessConsentResult();
-
-            var request = await _interactionService.GetAuthorizationContextAsync(model.ReturnUrl);
-            if (request is null)
-            {
-                return result;
-            }
-
-            ConsentResponse grantedConsent = null;
-            if (model?.Button == "no")
-            {
-                grantedConsent = new ConsentResponse
-                {
-                    Error = AuthorizationError.AccessDenied
-                };
-
-                await _eventService.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
-            }
-            else if (model?.Button == "yes")
-            {
-                if (model.ScopesConsented is not null && model.ScopesConsented.Any())
-                {
-                    var scopes = model.ScopesConsented;
-                    if (ConsentOptions.EnableOfflineAccess == false)
-                    {
-                        scopes = scopes.Where(x => x != IdentityServerConstants.StandardScopes.OfflineAccess);
-                    }
-
-                    grantedConsent = new ConsentResponse
-                    {
-                        RememberConsent = model.RememberConsent,
-                        ScopesValuesConsented = scopes.ToArray(),
-                        Description = model.Description
-                    };
-
-                    await _eventService.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
-                }
-                else
-                {
-                    result.ValidationError = ConsentOptions.MustChooseOneErrorMessage;
-                }
-            }
-            else
-            {
-                result.ValidationError = ConsentOptions.InvalidSelectionErrorMessage;
-            }
-
-            if (grantedConsent is not null)
-            {
-                await _interactionService.GrantConsentAsync(request, grantedConsent);
-
-                result.RedirectUri = model.ReturnUrl;
-                result.Client = request.Client;
-            }
-            else
-            {
-                result.ViewModel = await BuildViewModelAsync(model.ReturnUrl, model);
-            }
             return result;
         }
 
-        private async Task<ConsentViewModel> BuildViewModelAsync(string
-            returnUrl, ConsentInputModel model = null)
+        ConsentResponse grantedConsent = null;
+        if (model?.Button == "no")
         {
-            var request = await _interactionService.GetAuthorizationContextAsync(returnUrl);
-            if (request is not null)
+            grantedConsent = new ConsentResponse
             {
-                return CreateConsentViewModel(model, returnUrl, request);
+                Error = AuthorizationError.AccessDenied
+            };
+
+            await _eventService.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
+        }
+        else if (model?.Button == "yes")
+        {
+            if (model.ScopesConsented is not null && model.ScopesConsented.Any())
+            {
+                var scopes = model.ScopesConsented;
+                if (ConsentOptions.EnableOfflineAccess == false)
+                {
+                    scopes = scopes.Where(x => x != IdentityServerConstants.StandardScopes.OfflineAccess);
+                }
+
+                grantedConsent = new ConsentResponse
+                {
+                    RememberConsent = model.RememberConsent,
+                    ScopesValuesConsented = scopes.ToArray(),
+                    Description = model.Description
+                };
+
+                await _eventService.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
             }
             else
             {
-                _logger.LogError("No consent request matching request: {0}", returnUrl);
+                result.ValidationError = ConsentOptions.MustChooseOneErrorMessage;
             }
-            return null;
         }
-        private ConsentViewModel CreateConsentViewModel(ConsentInputModel model, string returnUrl, AuthorizationRequest request)
+        else
         {
-            var viewModel = new ConsentViewModel
+            result.ValidationError = ConsentOptions.InvalidSelectionErrorMessage;
+        }
+
+        if (grantedConsent is not null)
+        {
+            await _interactionService.GrantConsentAsync(request, grantedConsent);
+
+            result.RedirectUri = model.ReturnUrl;
+            result.Client = request.Client;
+        }
+        else
+        {
+            result.ViewModel = await BuildViewModelAsync(model.ReturnUrl, model);
+        }
+        return result;
+    }
+
+    private async Task<ConsentViewModel> BuildViewModelAsync(string
+        returnUrl, ConsentInputModel model = null)
+    {
+        var request = await _interactionService.GetAuthorizationContextAsync(returnUrl);
+        if (request is not null)
+        {
+            return CreateConsentViewModel(model, returnUrl, request);
+        }
+        else
+        {
+            _logger.LogError("No consent request matching request: {0}", returnUrl);
+        }
+        return null;
+    }
+    private ConsentViewModel CreateConsentViewModel(ConsentInputModel model, string returnUrl, AuthorizationRequest request)
+    {
+        var viewModel = new ConsentViewModel
+        {
+            RememberConsent = model?.RememberConsent ?? true,
+            ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
+            Description = model?.Description,
+            ReturnUrl = returnUrl,
+            ClientName = request.Client.ClientName ?? request.Client.ClientId,
+            ClientUrl = request.Client.ClientUri,
+            ClientLogoUrl = request.Client.LogoUri,
+            AllowRememberConsent = request.Client.AllowRememberConsent
+        };
+
+        viewModel.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, viewModel.ScopesConsented.Contains(x.Name) || model is null)).ToArray();
+
+        var apiScopes = new List<ScopeViewModel>();
+
+        foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
+        {
+            var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
+
+            if (apiScope is not null)
             {
-                RememberConsent = model?.RememberConsent ?? true,
-                ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
-                Description = model?.Description,
-                ReturnUrl = returnUrl,
-                ClientName = request.Client.ClientName ?? request.Client.ClientId,
-                ClientUrl = request.Client.ClientUri,
-                ClientLogoUrl = request.Client.LogoUri,
-                AllowRememberConsent = request.Client.AllowRememberConsent
-            };
+                var scopeViewModel = CreateScopeViewModel(parsedScope, apiScope, viewModel.ScopesConsented.Contains(parsedScope.RawValue) || model is null);
 
-            viewModel.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, viewModel.ScopesConsented.Contains(x.Name) || model is null)).ToArray();
-
-            var apiScopes = new List<ScopeViewModel>();
-
-            foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
-            {
-                var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
-
-                if (apiScope is not null)
-                {
-                    var scopeViewModel = CreateScopeViewModel(parsedScope, apiScope, viewModel.ScopesConsented.Contains(parsedScope.RawValue) || model is null);
-
-                    apiScopes.Add(scopeViewModel);
-                }
+                apiScopes.Add(scopeViewModel);
             }
-            if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
-            {
-                apiScopes.Add(GetOfflineAccessScope(viewModel.ScopesConsented.Contains(IdentityServerConstants.StandardScopes.OfflineAccess) || model is null));
-            }
+        }
+        if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
+        {
+            apiScopes.Add(GetOfflineAccessScope(viewModel.ScopesConsented.Contains(IdentityServerConstants.StandardScopes.OfflineAccess) || model is null));
+        }
 
-            viewModel.ApiScopes = apiScopes;
-            return viewModel;
-        }
-        private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+        viewModel.ApiScopes = apiScopes;
+        return viewModel;
+    }
+    private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+    {
+        return new ScopeViewModel
         {
-            return new ScopeViewModel
-            {
-                Value = identity.Name,
-                DisplayName = identity.DisplayName ?? identity.Name,
-                Description = identity.Description,
-                Emphasize = identity.Emphasize,
-                Required = identity.Required,
-                Checked = check || identity.Required
-            };
-        }
-        private ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+            Value = identity.Name,
+            DisplayName = identity.DisplayName ?? identity.Name,
+            Description = identity.Description,
+            Emphasize = identity.Emphasize,
+            Required = identity.Required,
+            Checked = check || identity.Required
+        };
+    }
+    private ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+    {
+        var displayName = apiScope.DisplayName ?? apiScope.Name;
+        if (!string.IsNullOrWhiteSpace(parsedScopeValue.ParsedParameter))
         {
-            var displayName = apiScope.DisplayName ?? apiScope.Name;
-            if (!string.IsNullOrWhiteSpace(parsedScopeValue.ParsedParameter))
-            {
-                displayName += ":" + parsedScopeValue.ParsedParameter;
-            }
+            displayName += ":" + parsedScopeValue.ParsedParameter;
+        }
 
-            return new ScopeViewModel
-            {
-                Value = parsedScopeValue.RawValue,
-                DisplayName = displayName,
-                Description = apiScope.Description,
-                Emphasize = apiScope.Emphasize,
-                Required = apiScope.Required,
-                Checked = check || apiScope.Required
-            };
-        }
-        private ScopeViewModel GetOfflineAccessScope(bool check)
+        return new ScopeViewModel
         {
-            return new ScopeViewModel
-            {
-                Value = IdentityServerConstants.StandardScopes.OfflineAccess,
-                DisplayName = ConsentOptions.OfflineAccessDisplayName,
-                Description = ConsentOptions.OfflineAccessDescrption,
-                Emphasize = true,
-                Checked = check
-            };
-        }
+            Value = parsedScopeValue.RawValue,
+            DisplayName = displayName,
+            Description = apiScope.Description,
+            Emphasize = apiScope.Emphasize,
+            Required = apiScope.Required,
+            Checked = check || apiScope.Required
+        };
+    }
+    private ScopeViewModel GetOfflineAccessScope(bool check)
+    {
+        return new ScopeViewModel
+        {
+            Value = IdentityServerConstants.StandardScopes.OfflineAccess,
+            DisplayName = ConsentOptions.OfflineAccessDisplayName,
+            Description = ConsentOptions.OfflineAccessDescrption,
+            Emphasize = true,
+            Checked = check
+        };
     }
 }
