@@ -1,241 +1,223 @@
-﻿using Identity.API.Models;
-using IdentityServer4;
-using IdentityServer4.Configuration;
-using IdentityServer4.Events;
-using IdentityServer4.Extensions;
-using IdentityServer4.Models;
-using IdentityServer4.Services;
-using IdentityServer4.Validation;
-using Microsoft.AspNetCore.Authorization;
-using Microsoft.AspNetCore.Mvc;
-using Microsoft.Extensions.Options;
-using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Threading.Tasks;
-
-namespace Identity.API.Controllers
+﻿namespace Identity.API.Controllers;
+[Authorize]
+public class DeviceController : Controller
 {
-    [Authorize]
-    public class DeviceController : Controller
+    private readonly IDeviceFlowInteractionService _interactionService;
+    private readonly IEventService _eventService;
+    private readonly IOptions<IdentityServerOptions> _options;
+
+    public DeviceController(
+        IDeviceFlowInteractionService interactionService,
+        IEventService eventService,
+        IOptions<IdentityServerOptions> options)
     {
-        private readonly IDeviceFlowInteractionService _interactionService;
-        private readonly IEventService _eventService;
-        private readonly IOptions<IdentityServerOptions> _options;
+        _interactionService = interactionService;
+        _eventService = eventService;
+        _options = options;
+    }
 
-        public DeviceController(
-            IDeviceFlowInteractionService interactionService,
-            IEventService eventService,
-            IOptions<IdentityServerOptions> options)
+    [HttpGet]
+    public async Task<IActionResult> Index()
+    {
+        string userCodeParamName = _options.Value.UserInteraction.DeviceVerificationUserCodeParameter;
+        string userCode = Request.Query[userCodeParamName];
+
+        if (string.IsNullOrWhiteSpace(userCode))
         {
-            _interactionService = interactionService;
-            _eventService = eventService;
-            _options = options;
+            return View("UserCodeCapture");
         }
 
-        [HttpGet]
-        public async Task<IActionResult> Index()
+        var model = await BuildViewModelAsync(userCode);
+
+        if (model is null)
         {
-            string userCodeParamName = _options.Value.UserInteraction.DeviceVerificationUserCodeParameter;
-            string userCode = Request.Query[userCodeParamName];
-
-            if (string.IsNullOrWhiteSpace(userCode))
-            {
-                return View("UserCodeCapture");
-            }
-
-            var model = await BuildViewModelAsync(userCode);
-
-            if (model is null)
-            {
-                return View("Error");
-            }
-
-            model.ConfirmUserCode = true;
-            return View("UserCodeConfirmation", model);
+            return View("Error");
         }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> UserCodeCapture(string userCode)
+        model.ConfirmUserCode = true;
+        return View("UserCodeConfirmation", model);
+    }
+
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> UserCodeCapture(string userCode)
+    {
+        var model = await BuildViewModelAsync(userCode);
+        if (model is null)
         {
-            var model = await BuildViewModelAsync(userCode);
-            if (model is null)
-            {
-                return View("Error");
-            }
-            return View("UserCodeConfirmation", model);
+            return View("Error");
         }
+        return View("UserCodeConfirmation", model);
+    }
 
-        [HttpPost]
-        [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Callback(DeviceAuthorizationInputModel model)
+    [HttpPost]
+    [ValidateAntiForgeryToken]
+    public async Task<IActionResult> Callback(DeviceAuthorizationInputModel model)
+    {
+        if (model is null)
         {
-            if (model is null)
-            {
-                throw new ArgumentNullException(nameof(model));
-            }
-            var result = await ProcessConsent(model);
-
-            if (result.HasValidationError)
-            {
-                return View("Error");
-            }
-            return View("Success");
+            throw new ArgumentNullException(nameof(model));
         }
+        var result = await ProcessConsent(model);
 
-
-        private async Task<ProcessConsentResult> ProcessConsent(DeviceAuthorizationInputModel model)
+        if (result.HasValidationError)
         {
-            var result = new ProcessConsentResult();
+            return View("Error");
+        }
+        return View("Success");
+    }
 
-            var request = await _interactionService.GetAuthorizationContextAsync(model.UserCode);
 
-            if (request is null)
-            {
-                return result;
-            }
+    private async Task<ProcessConsentResult> ProcessConsent(DeviceAuthorizationInputModel model)
+    {
+        var result = new ProcessConsentResult();
 
-            ConsentResponse grantedConsent = null;
+        var request = await _interactionService.GetAuthorizationContextAsync(model.UserCode);
 
-            if (model.Button == "no")
-            {
-                grantedConsent = new ConsentResponse
-                {
-                    Error = AuthorizationError.AccessDenied
-                };
-
-                await _eventService.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
-            }
-            else if (model.Button == "yes")
-            {
-                if (model.ScopesConsented is not null && model.ScopesConsented.Any())
-                {
-                    var scopes = model.ScopesConsented;
-                    if (ConsentOptions.EnableOfflineAccess == false)
-                    {
-                        scopes = scopes.Where(x => x != IdentityServerConstants.StandardScopes.OfflineAccess);
-                    }
-
-                    grantedConsent = new ConsentResponse
-                    {
-                        RememberConsent = model.RememberConsent,
-                        ScopesValuesConsented = scopes.ToArray(),
-                        Description = model.Description
-                    };
-
-                    await _eventService.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
-                }
-                else
-                {
-                    result.ValidationError = ConsentOptions.MustChooseOneErrorMessage;
-                }
-            }
-            else
-            {
-                result.ValidationError = ConsentOptions.InvalidSelectionErrorMessage;
-            }
-
-            if (grantedConsent is not null)
-            {
-                await _interactionService.HandleRequestAsync(model.UserCode, grantedConsent);
-
-                result.RedirectUri = model.ReturnUrl;
-                result.Client = request.Client;
-            }
-            else
-            {
-                result.ViewModel = await BuildViewModelAsync(model.UserCode, model);
-            }
-
+        if (request is null)
+        {
             return result;
         }
 
-        private async Task<DeviceAuthorizationViewModel> BuildViewModelAsync(string userCode, DeviceAuthorizationInputModel model = null)
+        ConsentResponse grantedConsent = null;
+
+        if (model.Button == "no")
         {
-            var request = await _interactionService.GetAuthorizationContextAsync(userCode);
-
-            if (request is not null)
+            grantedConsent = new ConsentResponse
             {
-                return CreateConsentViewModel(userCode, model, request);
-            }
-
-            return null;
-        }
-
-        private DeviceAuthorizationViewModel CreateConsentViewModel(string userCode, DeviceAuthorizationInputModel model, DeviceFlowAuthorizationRequest request)
-        {
-            var viewModel = new DeviceAuthorizationViewModel
-            {
-                UserCode = userCode,
-                Description = model?.Description,
-                RememberConsent = model?.RememberConsent ?? true,
-                ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
-                ClientName = request.Client.ClientName ?? request.Client.ClientId,
-                ClientUrl = request.Client.ClientUri,
-                ClientLogoUrl = request.Client.LogoUri,
-                AllowRememberConsent = request.Client.AllowRememberConsent
+                Error = AuthorizationError.AccessDenied
             };
 
-            viewModel.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, viewModel.ScopesConsented.Contains(x.Name) || model is null)).ToArray();
-
-            var apiScopes = new List<ScopeViewModel>();
-            foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
+            await _eventService.RaiseAsync(new ConsentDeniedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues));
+        }
+        else if (model.Button == "yes")
+        {
+            if (model.ScopesConsented is not null && model.ScopesConsented.Any())
             {
-                var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
-
-                if (apiScope is not null)
+                var scopes = model.ScopesConsented;
+                if (ConsentOptions.EnableOfflineAccess == false)
                 {
-                    var scopeViewMode = CreateScopeViewModel(parsedScope, apiScope, viewModel.ScopesConsented.Contains(parsedScope.RawValue) || model is null);
-
-                    apiScopes.Add(scopeViewMode);
+                    scopes = scopes.Where(x => x != IdentityServerConstants.StandardScopes.OfflineAccess);
                 }
+
+                grantedConsent = new ConsentResponse
+                {
+                    RememberConsent = model.RememberConsent,
+                    ScopesValuesConsented = scopes.ToArray(),
+                    Description = model.Description
+                };
+
+                await _eventService.RaiseAsync(new ConsentGrantedEvent(User.GetSubjectId(), request.Client.ClientId, request.ValidatedResources.RawScopeValues, grantedConsent.ScopesValuesConsented, grantedConsent.RememberConsent));
             }
-
-            if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
+            else
             {
-                apiScopes.Add(GetOfflineAccessScope(viewModel.ScopesConsented.Contains(IdentityServerConstants.StandardScopes.OfflineAccess) || model is null));
+                result.ValidationError = ConsentOptions.MustChooseOneErrorMessage;
             }
-
-            return viewModel;
         }
-
-        private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+        else
         {
-            return new ScopeViewModel
-            {
-                Value = identity.Name,
-                DisplayName = identity.DisplayName ?? identity.Name,
-                Description = identity.Description,
-                Emphasize = identity.Emphasize,
-                Required = identity.Required,
-                Checked = check || identity.Required
-            };
+            result.ValidationError = ConsentOptions.InvalidSelectionErrorMessage;
         }
 
-        private ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+        if (grantedConsent is not null)
         {
-            return new ScopeViewModel
-            {
-                Value = parsedScopeValue.RawValue,
-                DisplayName = apiScope.DisplayName ?? apiScope.Name,
-                Description = apiScope.Description,
-                Emphasize = apiScope.Emphasize,
-                Required = apiScope.Required,
-                Checked = check || apiScope.Required
-            };
+            await _interactionService.HandleRequestAsync(model.UserCode, grantedConsent);
+
+            result.RedirectUri = model.ReturnUrl;
+            result.Client = request.Client;
+        }
+        else
+        {
+            result.ViewModel = await BuildViewModelAsync(model.UserCode, model);
         }
 
-        private ScopeViewModel GetOfflineAccessScope(bool check)
+        return result;
+    }
+
+    private async Task<DeviceAuthorizationViewModel> BuildViewModelAsync(string userCode, DeviceAuthorizationInputModel model = null)
+    {
+        var request = await _interactionService.GetAuthorizationContextAsync(userCode);
+
+        if (request is not null)
         {
-            return new ScopeViewModel
-            {
-                Value = IdentityServerConstants.StandardScopes.OfflineAccess,
-                DisplayName = ConsentOptions.OfflineAccessDisplayName,
-                Description = ConsentOptions.OfflineAccessDescrption,
-                Emphasize = true,
-                Checked = check
-            };
+            return CreateConsentViewModel(userCode, model, request);
         }
+
+        return null;
+    }
+
+    private DeviceAuthorizationViewModel CreateConsentViewModel(string userCode, DeviceAuthorizationInputModel model, DeviceFlowAuthorizationRequest request)
+    {
+        var viewModel = new DeviceAuthorizationViewModel
+        {
+            UserCode = userCode,
+            Description = model?.Description,
+            RememberConsent = model?.RememberConsent ?? true,
+            ScopesConsented = model?.ScopesConsented ?? Enumerable.Empty<string>(),
+            ClientName = request.Client.ClientName ?? request.Client.ClientId,
+            ClientUrl = request.Client.ClientUri,
+            ClientLogoUrl = request.Client.LogoUri,
+            AllowRememberConsent = request.Client.AllowRememberConsent
+        };
+
+        viewModel.IdentityScopes = request.ValidatedResources.Resources.IdentityResources.Select(x => CreateScopeViewModel(x, viewModel.ScopesConsented.Contains(x.Name) || model is null)).ToArray();
+
+        var apiScopes = new List<ScopeViewModel>();
+        foreach (var parsedScope in request.ValidatedResources.ParsedScopes)
+        {
+            var apiScope = request.ValidatedResources.Resources.FindApiScope(parsedScope.ParsedName);
+
+            if (apiScope is not null)
+            {
+                var scopeViewMode = CreateScopeViewModel(parsedScope, apiScope, viewModel.ScopesConsented.Contains(parsedScope.RawValue) || model is null);
+
+                apiScopes.Add(scopeViewMode);
+            }
+        }
+
+        if (ConsentOptions.EnableOfflineAccess && request.ValidatedResources.Resources.OfflineAccess)
+        {
+            apiScopes.Add(GetOfflineAccessScope(viewModel.ScopesConsented.Contains(IdentityServerConstants.StandardScopes.OfflineAccess) || model is null));
+        }
+
+        return viewModel;
+    }
+
+    private ScopeViewModel CreateScopeViewModel(IdentityResource identity, bool check)
+    {
+        return new ScopeViewModel
+        {
+            Value = identity.Name,
+            DisplayName = identity.DisplayName ?? identity.Name,
+            Description = identity.Description,
+            Emphasize = identity.Emphasize,
+            Required = identity.Required,
+            Checked = check || identity.Required
+        };
+    }
+
+    private ScopeViewModel CreateScopeViewModel(ParsedScopeValue parsedScopeValue, ApiScope apiScope, bool check)
+    {
+        return new ScopeViewModel
+        {
+            Value = parsedScopeValue.RawValue,
+            DisplayName = apiScope.DisplayName ?? apiScope.Name,
+            Description = apiScope.Description,
+            Emphasize = apiScope.Emphasize,
+            Required = apiScope.Required,
+            Checked = check || apiScope.Required
+        };
+    }
+
+    private ScopeViewModel GetOfflineAccessScope(bool check)
+    {
+        return new ScopeViewModel
+        {
+            Value = IdentityServerConstants.StandardScopes.OfflineAccess,
+            DisplayName = ConsentOptions.OfflineAccessDisplayName,
+            Description = ConsentOptions.OfflineAccessDescrption,
+            Emphasize = true,
+            Checked = check
+        };
     }
 }
