@@ -1,15 +1,11 @@
-﻿using System.Collections.Generic;
-
-namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class ConsentViewModel : ConsentInputModel
 {
-    public class ConsentViewModel : ConsentInputModel
-    {
 
-        public string ClientName { get; set; }
-        public string ClientUrl { get; set; }
-        public string ClientLogoUrl { get; set; }
-        public bool AllowRememberConsent { get; set; }
-        public IEnumerable<ScopeViewModel> IdentityScopes { get; set; }
-        public IEnumerable<ScopeViewModel> ApiScopes { get; set; }
-    }
+    public string ClientName { get; set; }
+    public string ClientUrl { get; set; }
+    public string ClientLogoUrl { get; set; }
+    public bool AllowRememberConsent { get; set; }
+    public IEnumerable<ScopeViewModel> IdentityScopes { get; set; }
+    public IEnumerable<ScopeViewModel> ApiScopes { get; set; }
 }
