@@ -1,8 +1,6 @@
-﻿namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class DeviceAuthorizationViewModel : ConsentViewModel
 {
-    public class DeviceAuthorizationViewModel : ConsentViewModel
-    {
-        public string UserCode { get; set; }
-        public bool ConfirmUserCode { get; set; }
-    }
+    public string UserCode { get; set; }
+    public bool ConfirmUserCode { get; set; }
 }
