@@ -1,28 +1,20 @@
-﻿using IdentityModel;
-using Microsoft.AspNetCore.Authentication;
-using Newtonsoft.Json;
-using System.Collections.Generic;
-using System.Text;
-
-namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class DiagnosticsViewModel
 {
-    public class DiagnosticsViewModel
+    public AuthenticateResult AuthenticateResult { get; }
+    public IEnumerable<string> Clients { get; } = new List<string>();
+
+    public DiagnosticsViewModel(AuthenticateResult result)
     {
-        public AuthenticateResult AuthenticateResult { get; }
-        public IEnumerable<string> Clients { get; } = new List<string>();
+        AuthenticateResult = result;
 
-        public DiagnosticsViewModel(AuthenticateResult result)
+        if (result.Properties.Items.ContainsKey("client_list"))
         {
-            AuthenticateResult = result;
+            var encoded = result.Properties.Items["client_list"];
+            var bytes = Base64Url.Decode(encoded);
+            var value = Encoding.UTF8.GetString(bytes);
 
-            if (result.Properties.Items.ContainsKey("client_list"))
-            {
-                var encoded = result.Properties.Items["client_list"];
-                var bytes = Base64Url.Decode(encoded);
-                var value = Encoding.UTF8.GetString(bytes);
-
-                Clients = JsonConvert.DeserializeObject<string[]>(value);
-            }
+            Clients = JsonConvert.DeserializeObject<string[]>(value);
         }
     }
 }
