@@ -1,19 +1,15 @@
-﻿using IdentityServer4.Models;
-
-namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public record ErrorViewModel
 {
-    public record ErrorViewModel
+    public ErrorMessage Error { get; set; }
+
+    public ErrorViewModel()
     {
-        public ErrorMessage Error { get; set; }
 
-        public ErrorViewModel()
-        {
+    }
 
-        }
-
-        public ErrorViewModel(string error)
-        {
-            Error = new ErrorMessage { Error = error };
-        }
+    public ErrorViewModel(string error)
+    {
+        Error = new ErrorMessage { Error = error };
     }
 }
