@@ -1,8 +1,6 @@
-﻿namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class ExternalProvider
 {
-    public class ExternalProvider
-    {
-        public string DisplayName { get; set; }
-        public string AuthenticationScheme { get; set; }
-    }
+    public string DisplayName { get; set; }
+    public string AuthenticationScheme { get; set; }
 }
