@@ -1,7 +1,5 @@
-﻿namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class RedirectViewModel
 {
-    public class RedirectViewModel
-    {
-        public string RedirectUrl { get; set; }
-    }
+    public string RedirectUrl { get; set; }
 }
