@@ -1,25 +1,21 @@
-﻿using System.ComponentModel.DataAnnotations;
-
-namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class RegisterViewModel
 {
-    public class RegisterViewModel
-    {
-        [Required]
-        [EmailAddress]
-        [Display(Name = "Email")]
-        public string Email { get; init; }
+    [Required]
+    [EmailAddress]
+    [Display(Name = "Email")]
+    public string Email { get; init; }
 
-        [Required]
-        [StringLength(100, ErrorMessage = "The {0} must be at least {2} and at max {1} characters long.", MinimumLength = 6)]
-        [DataType(DataType.Password)]
-        [Display(Name = "Password")]
-        public string Password { get; init; }
+    [Required]
+    [StringLength(100, ErrorMessage = "The {0} must be at least {2} and at max {1} characters long.", MinimumLength = 6)]
+    [DataType(DataType.Password)]
+    [Display(Name = "Password")]
+    public string Password { get; init; }
 
-        [DataType(DataType.Password)]
-        [Display(Name = "Confirm Password")]
-        [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
-        public string ConfirmPassword { get; init; }
+    [DataType(DataType.Password)]
+    [Display(Name = "Confirm Password")]
+    [Compare("Password", ErrorMessage = "The password and confirmation password do not match.")]
+    public string ConfirmPassword { get; init; }
 
-        public ApplicationUser User { get; init; }
-    }
+    public ApplicationUser User { get; init; }
 }
