@@ -1,12 +1,10 @@
-﻿namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class ScopeViewModel
 {
-    public class ScopeViewModel
-    {
-        public string Value { get; set; }
-        public string DisplayName { get; set; }
-        public string Description { get; set; }
-        public bool Emphasize { get; set; }
-        public bool Required { get; set; }
-        public bool Checked { get; set; }
-    }
+    public string Value { get; set; }
+    public string DisplayName { get; set; }
+    public string Description { get; set; }
+    public bool Emphasize { get; set; }
+    public bool Required { get; set; }
+    public bool Checked { get; set; }
 }
