@@ -1,7 +1,5 @@
-﻿namespace Identity.API.Models
+﻿namespace Identity.API.Models;
+public class SignOutInputModel
 {
-    public class SignOutInputModel
-    {
-        public string SignOutId { get; set; }
-    }
+    public string SignOutId { get; set; }
 }
