@@ -1,39 +1,32 @@
-﻿using Identity.API.Models;
-using Microsoft.AspNetCore.Authentication;
-using Microsoft.AspNetCore.Identity;
-using System.Threading.Tasks;
-
-namespace Identity.API.Services
+﻿namespace Identity.API.Services;
+public class EFLoginService : ILoginService<ApplicationUser>
 {
-    public class EFLoginService : ILoginService<ApplicationUser>
+    private UserManager<ApplicationUser> _userManager;
+    private SignInManager<ApplicationUser> _signInManager;
+
+    public EFLoginService(UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signInManager)
     {
-        private UserManager<ApplicationUser> _userManager;
-        private SignInManager<ApplicationUser> _signInManager;
+        _userManager = userManager;
+        _signInManager = signInManager;
+    }
 
-        public EFLoginService(UserManager<ApplicationUser> userManager, SignInManager<ApplicationUser> signInManager)
-        {
-            _userManager = userManager;
-            _signInManager = signInManager;
-        }
+    public async Task<ApplicationUser> FindByUserName(string user)
+    {
+        return await _userManager.FindByNameAsync(user);
+    }
 
-        public async Task<ApplicationUser> FindByUserName(string user)
-        {
-            return await _userManager.FindByNameAsync(user);
-        }
+    public Task SignIn(ApplicationUser user)
+    {
+        return _signInManager.SignInAsync(user, true);
+    }
 
-        public Task SignIn(ApplicationUser user)
-        {
-            return _signInManager.SignInAsync(user, true);
-        }
+    public Task SignInAsync(ApplicationUser user, AuthenticationProperties properties, string authenticationMethod = null)
+    {
+        return _signInManager.SignInAsync(user, properties, authenticationMethod);
+    }
 
-        public Task SignInAsync(ApplicationUser user, AuthenticationProperties properties, string authenticationMethod = null)
-        {
-            return _signInManager.SignInAsync(user, properties, authenticationMethod);
-        }
-
-        public async Task<bool> ValidateCredentials(ApplicationUser user, string password)
-        {
-            return await _userManager.CheckPasswordAsync(user, password);
-        }
+    public async Task<bool> ValidateCredentials(ApplicationUser user, string password)
+    {
+        return await _userManager.CheckPasswordAsync(user, password);
     }
 }
